update files with working version
