+++ test/bf16.cpp
@@ -44,1094 +44,1045 @@ bool bit_equal(const T& x, const U& y)
 
 TEST_CASE(check_numeric_limits)
 {
-    std::bitset<8> exponentBits(std::numeric_limits<migraphx::bf16>::quiet_NaN().exponent);
-    std::cout << "Exponent bits: " << exponentBits << std::endl;
-
-    std::bitset<7> mantissaBits(std::numeric_limits<migraphx::bf16>::quiet_NaN().mantissa);
-    std::cout << "Mantissa bits: " << mantissaBits << std::endl;
-
-    std::cout << std::numeric_limits<migraphx::bf16>::min() << std::endl;
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::min(), uint16_t{0x0080}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::lowest(), uint16_t{0xff7f}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::max(), uint16_t{0x7f7f}));
-    CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::epsilon(), uint16_t{0x3c0}));
+    CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::epsilon(), uint16_t{0x3c00}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::denorm_min(), uint16_t{0x0001}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::infinity(), uint16_t{0x7f80}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::quiet_NaN(), uint16_t{0x7fc0}));
     CHECK(bit_equal(std::numeric_limits<migraphx::bf16>::signaling_NaN(), uint16_t{0x7fa0}));
 }
 
-
-
-
-static const std::map<uint16_t, float> half_lut = {
-{0x0000, 0.0},
-{0x004d, 0.0000000000000000000000000000000000000071},
-{0x0078, 0.0000000000000000000000000000000000000110},
-{0x00e4, 0.0000000000000000000000000000000000000209},
-{0x016c, 0.0000000000000000000000000000000000000433},
-{0x01da, 0.0000000000000000000000000000000000000801},
-{0x01dd, 0.0000000000000000000000000000000000000812},
-{0x0345, 0.0000000000000000000000000000000000005789},
-{0x0406, 0.0000000000000000000000000000000000015752},
-{0x0426, 0.0000000000000000000000000000000000019513},
-{0x04c2, 0.0000000000000000000000000000000000045609},
-{0x052f, 0.0000000000000000000000000000000000082285},
-{0x053f, 0.0000000000000000000000000000000000089808},
-{0x054c, 0.0000000000000000000000000000000000095920},
-{0x0648, 0.0000000000000000000000000000000000376158},
-{0x0659, 0.0000000000000000000000000000000000408132},
-{0x0676, 0.0000000000000000000000000000000000462675},
-{0x0695, 0.0000000000000000000000000000000000560476},
-{0x069a, 0.0000000000000000000000000000000000579284},
-{0x06b3, 0.0000000000000000000000000000000000673323},
-{0x06b9, 0.0000000000000000000000000000000000695893},
-{0x06cf, 0.0000000000000000000000000000000000778647},
-{0x0751, 0.0000000000000000000000000000000001572341},
-{0x0768, 0.0000000000000000000000000000000001745374},
-{0x076a, 0.0000000000000000000000000000000001760420},
-{0x07ba, 0.0000000000000000000000000000000002798617},
-{0x07c2, 0.0000000000000000000000000000000002918988},
-{0x07fe, 0.0000000000000000000000000000000003821767},
-{0x0801, 0.0000000000000000000000000000000003881953},
-{0x0838, 0.0000000000000000000000000000000005537049},
-{0x08fd, 0.0000000000000000000000000000000015226884},
-{0x0909, 0.0000000000000000000000000000000016490775},
-{0x0917, 0.0000000000000000000000000000000018175964},
-{0x0931, 0.000000000000000000000000000000002130560},
-{0x0953, 0.0000000000000000000000000000000025398201},
-{0x099f, 0.0000000000000000000000000000000038277858},
-{0x09b3, 0.0000000000000000000000000000000043092683},
-{0x09ec, 0.0000000000000000000000000000000056814933},
-{0x0a6a, 0.0000000000000000000000000000000112666902},
-{0x0ad7, 0.0000000000000000000000000000000207037469},
-{0x0ae2, 0.0000000000000000000000000000000217630084},
-{0x0b4c, 0.0000000000000000000000000000000392889709},
-{0x0c5b, 0.0000000000000000000000000000001687114631},
-{0x0cde, 0.0000000000000000000000000000003420451581},
-{0x0ce8, 0.0000000000000000000000000000003574525977},
-{0x0d3b, 0.0000000000000000000000000000005762382394},
-{0x0da1, 0.0000000000000000000000000000009922391073},
-{0x0dcc, 0.0000000000000000000000000000012572470677},
-{0x0e4d, 0.0000000000000000000000000000025268200870},
-{0x0e58, 0.0000000000000000000000000000026624055551},
-{0x0e94, 0.0000000000000000000000000000036484816866},
-{0x0eb8, 0.0000000000000000000000000000045359502050},
-{0x0efe, 0.0000000000000000000000000000062615834352},
-{0x0f61, 0.0000000000000000000000000000110933564797},
-{0x0f7b, 0.0000000000000000000000000000123752554507},
-{0x0ff0, 0.0000000000000000000000000000236658271566},
-{0x10f0, 0.0000000000000000000000000000946633086265},
-{0x10fe, 0.0000000000000000000000000001001853349631},
-{0x1109, 0.0000000000000000000000000001080739440153},
-{0x118a, 0.0000000000000000000000000002177256098410},
-{0x11c1, 0.0000000000000000000000000003045003094153},
-{0x11c3, 0.0000000000000000000000000003076557530362},
-{0x11d2, 0.0000000000000000000000000003313215801928},
-{0x1224, 0.0000000000000000000000000005174927538250},
-{0x12d7, 0.0000000000000000000000000013568407569801},
-{0x1306, 0.0000000000000000000000000016913177807938},
-{0x1309, 0.0000000000000000000000000017291831042445},
-{0x1351, 0.0000000000000000000000000026379508670591},
-{0x1366, 0.0000000000000000000000000029030081312133},
-{0x136c, 0.0000000000000000000000000029787387781145},
-{0x139c, 0.0000000000000000000000000039379936388633},
-{0x13a4, 0.0000000000000000000000000041399420305999},
-{0x13cb, 0.0000000000000000000000000051244404403157},
-{0x13f6, 0.0000000000000000000000000062099130458998},
-{0x13fe, 0.0000000000000000000000000064118614376364},
-{0x1423, 0.0000000000000000000000000082293969632656},
-{0x1471, 0.0000000000000000000000000121673906021289},
-{0x1566, 0.0000000000000000000000000464481300994132},
-{0x1571, 0.0000000000000000000000000486695624085155},
-{0x15ad, 0.0000000000000000000000000698741435408563},
-{0x15bf, 0.0000000000000000000000000771442856433732},
-{0x15d0, 0.0000000000000000000000000840105309624169},
-{0x15ef, 0.0000000000000000000000000965313312500848},
-{0x1659, 0.0000000000000000000000001752912040273506},
-{0x16d0, 0.0000000000000000000000003360421238496675},
-{0x173d, 0.0000000000000000000000006106919366114150},
-{0x174c, 0.0000000000000000000000006591595506281939},
-{0x1769, 0.0000000000000000000000007528636043939666},
-{0x17c4, 0.0000000000000000000000012666203129718236},
-{0x1882, 0.0000000000000000000000033604212384966749},
-{0x18ae, 0.0000000000000000000000044977945807570880},
-{0x18c1, 0.0000000000000000000000049889330694604482},
-{0x18d1, 0.0000000000000000000000054025233757369620},
-{0x18d2, 0.0000000000000000000000054283727698792441},
-{0x18e5, 0.0000000000000000000000059195112585826043},
-{0x1947, 0.0000000000000000000000102880588686282817},
-{0x1998, 0.0000000000000000000000157164316385075258},
-{0x19e9, 0.0000000000000000000000240916353406069310},
-{0x1a1b, 0.0000000000000000000000320532487364298224},
-{0x1a8a, 0.0000000000000000000000570754622661589096},
-{0x1a8c, 0.0000000000000000000000579026428787119372},
-{0x1b70, 0.0000000000000000000001985233470127266420},
-{0x1b90, 0.0000000000000000000002382280164152719704},
-{0x1bea, 0.0000000000000000000003871205266748169518},
-{0x1c1d, 0.0000000000000000000005194694246833013798},
-{0x1c7e, 0.0000000000000000000008404155023538761177},
-{0x1c91, 0.0000000000000000000009595295105615121029},
-{0x1cab, 0.0000000000000000000011315830779725418592},
-{0x1d62, 0.0000000000000000000029910850949917480723},
-{0x1e7d, 0.0000000000000000000133937084784586241115},
-{0x1eca, 0.0000000000000000000213875819181710835615},
-{0x1efb, 0.0000000000000000000265756587201036731383},
-{0x1efc, 0.0000000000000000000266815378385104606807},
-{0x1f0f, 0.0000000000000000000302814278643412371217},
-{0x1f6b, 0.0000000000000000000497631856511901449203},
-{0x1f85, 0.0000000000000000000563276909924109725480},
-{0x1f8a, 0.0000000000000000000584452733605467233957},
-{0x1f9c, 0.0000000000000000000660685698858354264473},
-{0x1fdc, 0.0000000000000000000931736241979730372975},
-{0x1ffc, 0.0000000000000000001067261513540418427226},
-{0x2084, 0.0000000000000000002236166980751352895140},
-{0x20fb, 0.0000000000000000004252105395216587702123},
-{0x2153, 0.0000000000000000007148958074826294861737},
-{0x21bb, 0.0000000000000000012671612890924333072462},
-{0x2207, 0.0000000000000000018295911660692887323876},
-{0x220c, 0.0000000000000000018973538018496327595130},
-{0x2265, 0.0000000000000000031035287187397564423463},
-{0x2268, 0.0000000000000000031441863002079628586216},
-{0x2298, 0.0000000000000000041199682554449168492283},
-{0x22a1, 0.000000000000000004363913744254155346880},
-{0x2310, 0.0000000000000000078062556418956319248537},
-{0x2313, 0.0000000000000000079688859677684575899548},
-{0x232d, 0.0000000000000000093783487919996133541645},
-{0x23e5, 0.0000000000000000248282297499180515387707},
-{0x23f5, 0.0000000000000000265629532258948586331826},
-{0x2452, 0.0000000000000000455364912443911862283130},
-{0x2462, 0.0000000000000000490059381963448004171369},
-{0x2472, 0.0000000000000000524753851482984146059607},
-{0x24be, 0.0000000000000000823993651088983369845664},
-{0x24e3, 0.0000000000000000984455572616838026078767},
-{0x2509, 0.0000000000000001188285581044112859672168},
-{0x250b, 0.0000000000000001205632815803880930616288},
-{0x2546, 0.0000000000000001717376241217039023467805},
-{0x254a, 0.0000000000000001752070710736575165356044},
-{0x2550, 0.0000000000000001804112415015879378188401},
-{0x2585, 0.0000000000000002307182223049153435567860},
-{0x25d2, 0.0000000000000003642919299551294898265041},
-{0x2608, 0.0000000000000004718447854656915296800435},
-{0x261d, 0.0000000000000005447031714567174276453443},
-{0x2622, 0.0000000000000005620504062164854985894635},
-{0x26e1, 0.0000000000000015612511283791263849707320},
-{0x2726, 0.0000000000000023037127760971998213790357},
-{0x2731, 0.0000000000000024563684419831588456872851},
-{0x2737, 0.0000000000000025396351688300455862190574},
-{0x2750, 0.0000000000000028865798640254070051014423},
-{0x275c, 0.0000000000000030531133177191804861649871},
-{0x2776, 0.0000000000000034139358007223563618026674},
-{0x2781, 0.0000000000000035804692544161298428662121},
-{0x27c3, 0.0000000000000054123372450476381345652044},
-{0x27e3, 0.0000000000000063005156647477633669041097},
-{0x2804, 0.0000000000000073274719625260331667959690},
-{0x2838, 0.0000000000000102140518265514401718974113},
-{0x2887, 0.0000000000000149880108324396132957190275},
-{0x28e0, 0.0000000000000248689957516035065054893494},
-{0x28fe, 0.0000000000000281996648254789761267602444},
-{0x29b4, 0.0000000000000799360577730112709105014801},
-{0x29d6, 0.0000000000000950350909079133998602628708},
-{0x29f9, 0.0000000000001105782132526655914261937141},
-{0x2a58, 0.0000000000001918465386552270501852035522},
-{0x2aac, 0.0000000000003055333763768430799245834351},
-{0x2aea, 0.0000000000004156675004196586087346076965},
-{0x2b19, 0.0000000000005435651928564766421914100647},
-{0x2b80, 0.0000000000009094947017729282379150390625},
-{0x2bff, 0.0000000000018118839761882554739713668823},
-{0x2c4c, 0.0000000000028990143619012087583541870117},
-{0x2c8d, 0.0000000000040074610296869650483131408691},
-{0x2ccb, 0.0000000000057696070143720135092735290527},
-{0x2cfd, 0.0000000000071906924858922138810157775879},
-{0x2d31, 0.0000000000100612851383630186319351196289},
-{0x2d68, 0.0000000000131876731757074594497680664062},
-{0x2d77, 0.0000000000140403244586195796728134155273},
-{0x2dab, 0.0000000000194404492503963410854339599609},
-{0x2db1, 0.0000000000201225702767260372638702392578},
-{0x2de3, 0.0000000000258069121628068387508392333984},
-{0x2df1, 0.0000000000273985278909094631671905517578},
-{0x2e6a, 0.0000000000532054400537163019180297851562},
-{0x2eaa, 0.0000000000773070496506989002227783203125},
-{0x2ecf, 0.0000000000941327016334980726242065429688},
-{0x2f88, 0.0000000002473825588822364807128906250},
-{0x2fbb, 0.0000000003401510184630751609802246093750},
-{0x2fcd, 0.0000000003728928277269005775451660156250},
-{0x3035, 0.000000000658474164083600044250488281250},
-{0x3065, 0.000000000833097146824002265930175781250},
-{0x30ad, 0.00000000125874066725373268127441406250},
-{0x30ea, 0.0000000017025740817189216613769531250},
-{0x3143, 0.0000000028376234695315361022949218750},
-{0x317d, 0.0000000036816345527768135070800781250},
-{0x31eb, 0.000000006839400157332420349121093750},
-{0x31fa, 0.00000000727595761418342590332031250},
-{0x325b, 0.00000001274747774004936218261718750},
-{0x32da, 0.000000025378540158271789550781250},
-{0x332e, 0.00000004051253199577331542968750},
-{0x3333, 0.000000041676685214042663574218750},
-{0x336e, 0.00000005541369318962097167968750},
-{0x33aa, 0.0000000791624188423156738281250},
-{0x3446, 0.000000184401869773864746093750},
-{0x3456, 0.000000199303030967712402343750},
-{0x345c, 0.00000020489096641540527343750},
-{0x345d, 0.0000002058222889900207519531250},
-{0x350e, 0.0000005289912223815917968750},
-{0x3551, 0.00000077858567237854003906250},
-{0x357c, 0.000000938773155212402343750},
-{0x3684, 0.000003933906555175781250},
-{0x369f, 0.00000473856925964355468750},
-{0x36b2, 0.0000053048133850097656250},
-{0x37ab, 0.000020384788513183593750},
-{0x37c1, 0.000023007392883300781250},
-{0x37e4, 0.0000271797180175781250},
-{0x3814, 0.000035285949707031250},
-{0x3863, 0.00005412101745605468750},
-{0x3869, 0.00005555152893066406250},
-{0x388c, 0.00006675720214843750},
-{0x3895, 0.0000710487365722656250},
-{0x38ab, 0.0000815391540527343750},
-{0x38d0, 0.000099182128906250},
-{0x38e6, 0.000109672546386718750},
-{0x3930, 0.00016784667968750},
-{0x393a, 0.00017738342285156250},
-{0x395f, 0.000212669372558593750},
-{0x3978, 0.000236511230468750},
-{0x397a, 0.00023841857910156250},
-{0x3a1a, 0.000587463378906250},
-{0x3aac, 0.0013122558593750},
-{0x3af3, 0.001853942871093750},
-{0x3bcf, 0.0063171386718750},
-{0x3c1d, 0.009582519531250},
-{0x3c26, 0.01013183593750},
-{0x3c2a, 0.01037597656250},
-{0x3c71, 0.014709472656250},
-{0x3c7a, 0.01525878906250},
-{0x3cef, 0.02917480468750},
-{0x3d0c, 0.03417968750},
-{0x3d5e, 0.054199218750},
-{0x3d61, 0.0549316406250},
-{0x3e47, 0.19433593750},
-{0x3e96, 0.292968750},
-{0x3e9d, 0.3066406250},
-{0x3eb1, 0.3457031250},
-{0x3ef9, 0.4863281250},
-{0x3f7b, 0.980468750},
-{0x3f7d, 0.988281250},
-{0x3fb7, 1.42968750},
-{0x3ff0, 1.8750},
-{0x402d, 2.7031250},
-{0x4042, 3.031250},
-{0x405d, 3.4531250},
-{0x40a6, 5.18750},
-{0x40b1, 5.531250},
-{0x40b9, 5.781250},
-{0x40f2, 7.56250},
-{0x4128, 10.50},
-{0x416d, 14.81250},
-{0x41ef, 29.8750},
-{0x41f5, 30.6250},
-{0x4204, 33.0},
-{0x4237, 45.750},
-{0x429b, 77.50},
-{0x429d, 78.50},
-{0x42ee, 119.0},
-{0x43b4, 360.0},
-{0x43e7, 462.0},
-{0x43f7, 494.0},
-{0x4469, 932.0},
-{0x446d, 948.0},
-{0x447a, 1000.0},
-{0x447f, 1020.0},
-{0x44ca, 1616.0},
-{0x4525, 2640.0},
-{0x458b, 4448.0},
-{0x45a2, 5184.0},
-{0x464b, 12992.0},
-{0x47a0, 81920.0},
-{0x47a2, 82944.0},
-{0x480e, 145408.0},
-{0x4983, 1073152.0},
-{0x49bb, 1531904.0},
-{0x49f4, 1998848.0},
-{0x49fd, 2072576.0},
-{0x4a50, 3407872.0},
-{0x4a72, 3964928.0},
-{0x4aa4, 5373952.0},
-{0x4af1, 7897088.0},
-{0x4b04, 8650752.0},
-{0x4b12, 9568256.0},
-{0x4b13, 9633792.0},
-{0x4b6f, 15663104.0},
-{0x4c12, 38273024.0},
-{0x4c24, 42991616.0},
-{0x4c36, 47710208.0},
-{0x4c46, 51904512.0},
-{0x4c50, 54525952.0},
-{0x4cf0, 125829120.0},
-{0x4d8d, 295698432.0},
-{0x4dcb, 425721856.0},
-{0x4e25, 692060160.0},
-{0x4e91, 1216348160.0},
-{0x4ed0, 1744830464.0},
-{0x4ef3, 2038431744.0},
-{0x4efc, 2113929216.0},
-{0x4f43, 3271557120.0},
-{0x4fac, 5771362304.0},
-{0x4fc2, 6509559808.0},
-{0x4fea, 7851737088.0},
-{0x50a1, 21609054208.0},
-{0x50e1, 30198988800.0},
-{0x50f6, 33017561088.0},
-{0x51ae, 93415538688.0},
-{0x5205, 142807662592.0},
-{0x5218, 163208757248.0},
-{0x521c, 167503724544.0},
-{0x5362, 970662608896.0},
-{0x546f, 4105988734976.0},
-{0x5476, 4226247819264.0},
-{0x54d2, 7215545057280.0},
-{0x5511, 9964324126720.0},
-{0x569d, 86311662780416.0},
-{0x56f9, 136889197658112.0},
-{0x5702, 142936511610880.0},
-{0x570b, 152832116260864.0},
-{0x573a, 204509162766336.0},
-{0x578f, 314460325543936.0},
-{0x5812, 642114790621184.0},
-{0x584e, 905997581287424.0},
-{0x5851, 919191720820736.0},
-{0x5860, 985162418487296.0},
-{0x5907, 2374945115996160.0},
-{0x5938, 3236962232172544.0},
-{0x59e3, 7986852464164864.0},
-{0x5a27, 11751580277669888.0},
-{0x5a5b, 15410754974908416.0},
-{0x5aad, 24347585485471744.0},
-{0x5ab7, 25754960369025024.0},
-{0x5ada, 30680772461461504.0},
-{0x5aef, 33636259716923392.0},
-{0x5b09, 38562071809359872.0},
-{0x5c1f, 179018085187977216.0},
-{0x5d62, 1017813515785732096.0},
-{0x5d70, 1080863910568919040.0},
-{0x5dbb, 1684346260636565504.0},
-{0x5dea, 2107684625609392128.0},
-{0x5e0a, 2485986994308513792.0},
-{0x5e2e, 3134505340649865216.0},
-{0x5e3d, 3404721318292094976.0},
-{0x5ea7, 6016809102166982656.0},
-{0x5f1f, 11457157452030541824.0},
-{0x5f51, 15060037153926938624.0},
-{0x5f64, 16429131440647569408.0},
-{0x5f99, 22049623775605948416.0},
-{0x5fda, 31417111000536580096.0},
-{0x5ffc, 36317027395115679744.0},
-{0x604a, 58222535982645772288.0},
-{0x607f, 73498745918686494720.0},
-{0x608a, 79551583817872441344.0},
-{0x60c1, 111256925194560733184.0},
-{0x60eb, 135468276791304519680.0},
-{0x614b, 234043065435189936128.0},
-{0x616c, 272089475087215886336.0},
-{0x61cc, 470391973879593566208.0},
-{0x6270, 1106804644422573096960.0},
-{0x6291, 1337388945343942492160.0},
-{0x62bd, 1743217314965552627712.0},
-{0x630c, 2582544170319337226240.0},
-{0x6325, 3043712772162076016640.0},
-{0x633f, 3523328118078524358656.0},
-{0x63a0, 5902958103587056517120.0},
-{0x63b3, 6603934378388019478528.0},
-{0x63e0, 8264141345021879123968.0},
-{0x63eb, 8669969714643489259520.0},
-{0x63fc, 9297159013149614014464.0},
-{0x6427, 12322425041237980479488.0},
-{0x6469, 17192365476697302106112.0},
-{0x648c, 20660353362554697809920.0},
-{0x64b3, 26415737513552077914112.0},
-{0x64c3, 28776920754986900520960.0},
-{0x64ea, 34532304905984280625152.0},
-{0x64ed, 34975026763753309863936.0},
-{0x6503, 38664375578495220187136.0},
-{0x6531, 52241179216745450176512.0},
-{0x6566, 67884018191251149946880.0},
-{0x657b, 74082124200017559289856.0},
-{0x65af, 103301766812773489049600.0},
-{0x65b5, 106843541674925722959872.0},
-{0x65d9, 128094190847839126421504.0},
-{0x6611, 171185785004024638996480.0},
-{0x6613, 173546968245459461603328.0},
-{0x6635, 213687083349851445919744.0},
-{0x6699, 361261035939527858847744.0},
-{0x66c0, 453347182355485940514816.0},
-{0x6768, 1095589024025757689577472.0},
-{0x6798, 1435599410792372144963584.0},
-{0x67d3, 1992838655770990280179712.0},
-{0x67d6, 2021172854668208151461888.0},
-{0x681d, 2965646151242137194201088.0},
-{0x6854, 4004566777473459141214208.0},
-{0x6857, 4061235175267894883778560.0},
-{0x6876, 4646808619143730890276864.0},
-{0x689b, 5855734438758360064983040.0},
-{0x68ed, 8953606851520847325167616.0},
-{0x68fc, 9520290829465204750811136.0},
-{0x6908, 10275869466724347985002496.0},
-{0x6aaa, 102758694667243479850024960.0},
-{0x6ad0, 125728285239921434169442304.0},
-{0x6b26, 200681686056028443001225216.0},
-{0x6b60, 270799383593676935134183424.0},
-{0x6bd1, 505330992598914995027181568.0},
-{0x6c4b, 981647765527078889861414912.0},
-{0x6cb4, 1740853180245066011576893440.0},
-{0x6d68, 4487532642409503496509325312.0},
-{0x6d86, 5183873914507529901140082688.0},
-{0x6ead, 26770453349546348444693561344.0},
-{0x6f07, 41780476325881584277845442560.0},
-{0x6f1e, 48898631551772520858515406848.0},
-{0x6f69, 72110007288373401012873986048.0},
-{0x6fb5, 112033573555326914878370742272.0},
-{0x70a0, 396140812571321687967719751680.0},
-{0x713a, 921027389228322924524948422656.0},
-{0x713b, 925979149385464445624544919552.0},
-{0x715c, 1089387234571134641911229317120.0},
-{0x7178, 1228036518971097232699931230208.0},
-{0x719a, 1525142128399588498675721043968.0},
-{0x719e, 1564756209656720667472493019136.0},
-{0x71a5, 1634080851856701962866843975680.0},
-{0x71c4, 1941089981599476271041826783232.0},
-{0x71e0, 2218388550399401452619230609408.0},
-{0x721b, 3070091297427743081749828075520.0},
-{0x7333, 14181841090053316429244367110144.0},
-{0x734c, 16162545152909924869082965868544.0},
-{0x7354, 16796370453024039569831317471232.0},
-{0x7376, 19490127978509027048011811782656.0},
-{0x73fc, 39930993907189226147146150969344.0},
-{0x745d, 70037695662609674432692852097024.0},
-{0x74f1, 152751897327501642880352736247808.0},
-{0x7538, 233247710441994209875393389789184.0},
-{0x7539, 234515361042222439276890092994560.0},
-{0x7562, 286489035651579844738254924414976.0},
-{0x7563, 287756686251808074139751627620352.0},
-{0x759f, 403112890872576949675951619309568.0},
-{0x760d, 714954938528721382444140607832064.0},
-{0x768e, 1440051081859268600100254841307136.0},
-{0x7691, 1470474696264746105736175718236160.0},
-{0x770c, 2839537344511233859352615180042240.0},
-{0x771c, 3164055898169660586135771200618496.0},
-{0x776a, 4746083847254490879203656800927744.0},
-{0x7773, 4928625533687355913019182062501888.0},
-{0x7798, 6165852519510107808879964390948864.0},
-{0x77ca, 8194093479875274851274689519550464.0},
-{0x77db, 8883695406399431645688896063275008.0},
-{0x77eb, 9532732513716285099255208104427520.0},
-{0x7855, 17280612982311223201203058095685632.0},
-{0x788c, 22716298756089870874820921440337920.0},
-{0x789a, 24987928631698857962303013584371712.0},
-{0x789b, 25150187908528071325694591594659840.0},
-{0x78de, 36021559456085366672930318283964416.0},
-{0x7900, 41538374868278621028243970633760768.0},
-{0x7915, 48353264495105582290690247065862144.0},
-{0x791a, 49975857263397715924606027168743424.0},
-{0x7999, 99302677419478578395645742296334336.0},
-{0x79c6, 128509347248736983806129784148197376.0},
-{0x79d9, 140841052287757199423889712930095104.0},
-{0x7a28, 218076468058462760398280845827244032.0},
-{0x7a2d, 224566839131631294933943966238769152.0},
-{0x7a4f, 268701362429177329776453185037139968.0},
-{0x7aa2, 420576045541321037910970202666827776.0},
-{0x7aa9, 438749084546192934610826939819098112.0},
-{0x7b14, 768459935063154489022513456724574208.0},
-{0x7b2e, 903459653385060007364306361284296704.0},
-{0x7b4b, 1054036262282570008591690754831679488.0},
-{0x7b93, 1526535276409239322787965920790708224.0},
-{0x7b99, 1588842838711657254330331876741349376.0},
-{0x7bbe, 1973072806243234498841588605103636480.0},
-{0x7bc0, 1993841993677373809355710590420516864.0},
-{0x7c49, 4174606674262001413338519048692957184.0},
-{0x7c7f, 5296142795705524181101106255804497920.0},
-{0x7c98, 6313832979978350396293083536331636736.0},
-{0x7cda, 9055365721284739384157185598159847424.0},
-{0x7d06, 11132284464698670435569384129847885824.0},
-{0x7d1d, 13043049708639487002868606779000881152.0},
-{0x7d3d, 15701505700209318748676220899561570304.0},
-{0x7d4c, 16947656946257677379523540018574393344.0},
-{0x7d58, 17944577943096364284201395313784651776.0},
-{0x7d85, 22098415429924226387025792377160728576.0},
-{0x7da9, 28079941410956347815092924148422279168.0},
-{0x7db1, 29409169406741263687996731208702623744.0},
-{0x7db2, 29575322906214378172109707091237666816.0},
-{0x7dfc, 41870681867224849996469922398830854144.0},
-{0x7e0f, 47519900849310742456311102405022318592.0},
-{0x7e15, 49513742842988116265666812995442835456.0},
-{0x7e50, 69119855780815625390997967134577917952.0},
-{0x7e55, 70781390775546770232127725959928348672.0},
-{0x7e73, 80750600743933639278906278912030932992.0},
-{0x7f08, 180775007426748558714917760198126862336.0},
-{0x7f67, 307051667026315566640779430924759597056.0},
-{0x7f6a, 311039351013670314259490852105600630784.0},
-{0x7f7b, 333636226942013884098855572130366488576.0},
-{0x7f7e, 337623910929368631717566993311207522304.0},
-{0x7f9a, std::numeric_limits<float>::quiet_NaN()},
-{0x8022, -0.0000000000000000000000000000000000000031},
-{0x8075, -0.0000000000000000000000000000000000000107},
-{0x807d, -0.0000000000000000000000000000000000000115},
-{0x8095, -0.0000000000000000000000000000000000000137},
-{0x80c6, -0.0000000000000000000000000000000000000182},
-{0x80fc, -0.0000000000000000000000000000000000000231},
-{0x814d, -0.0000000000000000000000000000000000000377},
-{0x8181, -0.0000000000000000000000000000000000000474},
-{0x818a, -0.0000000000000000000000000000000000000507},
-{0x819c, -0.0000000000000000000000000000000000000573},
-{0x81be, -0.0000000000000000000000000000000000000698},
-{0x8201, -0.0000000000000000000000000000000000000948},
-{0x8232, -0.0000000000000000000000000000000000001308},
-{0x8241, -0.0000000000000000000000000000000000001418},
-{0x82c3, -0.0000000000000000000000000000000000002865},
-{0x82f2, -0.0000000000000000000000000000000000003556},
-{0x82fd, -0.0000000000000000000000000000000000003718},
-{0x8340, -0.0000000000000000000000000000000000005642},
-{0x8349, -0.0000000000000000000000000000000000005907},
-{0x835b, -0.0000000000000000000000000000000000006436},
-{0x83a0, -0.0000000000000000000000000000000000009404},
-{0x83b6, -0.0000000000000000000000000000000000010697},
-{0x83ca, -0.0000000000000000000000000000000000011872},
-{0x842e, -0.0000000000000000000000000000000000020454},
-{0x8445, -0.0000000000000000000000000000000000023157},
-{0x847d, -0.0000000000000000000000000000000000029740},
-{0x8486, -0.0000000000000000000000000000000000031503},
-{0x84bc, -0.0000000000000000000000000000000000044199},
-{0x84da, -0.0000000000000000000000000000000000051252},
-{0x84e2, -0.0000000000000000000000000000000000053132},
-{0x84fd, -0.0000000000000000000000000000000000059480},
-{0x8546, -0.0000000000000000000000000000000000093099},
-{0x8594, -0.0000000000000000000000000000000000139179},
-{0x859a, -0.0000000000000000000000000000000000144821},
-{0x85e3, -0.0000000000000000000000000000000000213470},
-{0x85e5, -0.0000000000000000000000000000000000215351},
-{0x86ee, -0.0000000000000000000000000000000000895256},
-{0x86ef, -0.0000000000000000000000000000000000899018},
-{0x871d, -0.0000000000000000000000000000000001181137},
-{0x876b, -0.0000000000000000000000000000000001767944},
-{0x8783, -0.0000000000000000000000000000000001971069},
-{0x87a9, -0.0000000000000000000000000000000002542829},
-{0x87fb, -0.0000000000000000000000000000000003776628},
-{0x8822, -0.0000000000000000000000000000000004875010},
-{0x885e, -0.0000000000000000000000000000000006680569},
-{0x888a, -0.0000000000000000000000000000000008305573},
-{0x8891, -0.0000000000000000000000000000000008726870},
-{0x891c, -0.0000000000000000000000000000000018777817},
-{0x89a0, -0.0000000000000000000000000000000038518599},
-{0x89d0, -0.0000000000000000000000000000000050074179},
-{0x8a05, -0.0000000000000000000000000000000064037171},
-{0x8a3a, -0.0000000000000000000000000000000089555742},
-{0x8a3f, -0.0000000000000000000000000000000091963155},
-{0x8a7e, -0.0000000000000000000000000000000122296551},
-{0x8ae4, -0.0000000000000000000000000000000219556014},
-{0x8afb, -0.0000000000000000000000000000000241704208},
-{0x8b18, -0.0000000000000000000000000000000292741352},
-{0x8b86, -0.0000000000000000000000000000000516149225},
-{0x8ba2, -0.0000000000000000000000000000000624001302},
-{0x8be0, -0.0000000000000000000000000000000862816615},
-{0x8c00, -0.0000000000000000000000000000000986076132},
-{0x8c34, -0.0000000000000000000000000000001386669560},
-{0x8c9f, -0.0000000000000000000000000000002449782889},
-{0x8caa, -0.0000000000000000000000000000002619264724},
-{0x8cb3, -0.0000000000000000000000000000002757931680},
-{0x8d34, -0.0000000000000000000000000000005546678240},
-{0x8d48, -0.0000000000000000000000000000006162975822},
-{0x8d58, -0.0000000000000000000000000000006656013888},
-{0x8d5f, -0.0000000000000000000000000000006871718042},
-{0x8d6a, -0.0000000000000000000000000000007210681712},
-{0x8d7f, -0.0000000000000000000000000000007857794173},
-{0x8d91, -0.0000000000000000000000000000008936314942},
-{0x8df0, -0.0000000000000000000000000000014791141973},
-{0x8e06, -0.0000000000000000000000000000016516775203},
-{0x8e1a, -0.0000000000000000000000000000018981965532},
-{0x8e9f, -0.0000000000000000000000000000039196526228},
-{0x8ed1, -0.0000000000000000000000000000051522477872},
-{0x8f1d, -0.0000000000000000000000000000077406976325},
-{0x8f23, -0.0000000000000000000000000000080365204719},
-{0x8f4f, -0.0000000000000000000000000000102058879613},
-{0x901b, -0.0000000000000000000000000000305683600773},
-{0x9059, -0.0000000000000000000000000000427957041082},
-{0x90c1, -0.0000000000000000000000000000761250773538},
-{0x911c, -0.0000000000000000000000000001230623012145},
-{0x9131, -0.0000000000000000000000000001396283802241},
-{0x919d, -0.0000000000000000000000000002477023242394},
-{0x9243, -0.0000000000000000000000000006153115060724},
-{0x9268, -0.0000000000000000000000000007320629200451},
-{0x92b7, -0.0000000000000000000000000011548923652436},
-{0x9308, -0.0000000000000000000000000017165613297609},
-{0x9319, -0.0000000000000000000000000019311314959810},
-{0x9353, -0.0000000000000000000000000026631944160261},
-{0x9411, -0.0000000000000000000000000073206292004510},
-{0x941e, -0.0000000000000000000000000079769614735949},
-{0x943d, -0.0000000000000000000000000095420615095534},
-{0x9456, -0.0000000000000000000000000108042389579070},
-{0x947d, -0.0000000000000000000000000127732357773386},
-{0x94f8, -0.0000000000000000000000000250416005753358},
-{0x95b2, -0.0000000000000000000000000718936274582221},
-{0x9629, -0.0000000000000000000000001365171128139274},
-{0x965a, -0.0000000000000000000000001760989975942969},
-{0x966c, -0.0000000000000000000000001906392817993306},
-{0x971a, -0.0000000000000000000000004976008372389307},
-{0x9728, -0.0000000000000000000000005428372769879244},
-{0x977b, -0.0000000000000000000000008110247412141014},
-{0x97b9, -0.0000000000000000000000011955344790805478},
-{0x98f2, -0.0000000000000000000000062555533824322718},
-{0x9912, -0.0000000000000000000000075480230895463775},
-{0x9973, -0.0000000000000000000000125628055531491078},
-{0x9978, -0.0000000000000000000000128212994945719290},
-{0x9994, -0.0000000000000000000000153028413322310120},
-{0x9a97, -0.0000000000000000000000624521362477535895},
-{0x9aa0, -0.0000000000000000000000661744490042422140},
-{0x9aaf, -0.0000000000000000000000723783035983899216},
-{0x9ac4, -0.0000000000000000000000810637000301967121},
-{0x9ad9, -0.0000000000000000000000897490964620035027},
-{0x9b02, -0.0000000000000000000001075334796318935977},
-{0x9b19, -0.0000000000000000000001265586337206132343},
-{0x9b75, -0.0000000000000000000002026592500754917803},
-{0x9ba8, -0.0000000000000000000002779326858178172988},
-{0x9bb5, -0.0000000000000000000002994393817441960183},
-{0x9bc1, -0.0000000000000000000003192917164454686825},
-{0x9be2, -0.0000000000000000000003738856368739685090},
-{0x9c65, -0.0000000000000000000007576974410985733502},
-{0x9c76, -0.0000000000000000000008139457227521792321},
-{0x9cb0, -0.0000000000000000000011646703024746629662},
-{0x9cef, -0.0000000000000000000015815693312013889144},
-{0x9d01, -0.0000000000000000000017073007843094491209},
-{0x9d33, -0.0000000000000000000023690452743518712608},
-{0x9d91, -0.0000000000000000000038381180422460484114},
-{0x9db1, -0.0000000000000000000046851509895003487505},
-{0x9de7, -0.0000000000000000000061145190879919805726},
-{0x9dea, -0.0000000000000000000061939284267970712294},
-{0x9e48, -0.0000000000000000000105879118406787542384},
-{0x9fb7, -0.0000000000000000000775035146737684810248},
-{0x9ff7, -0.0000000000000000001046085689859060918749},
-{0xa009, -0.0000000000000000001160435137738391464524},
-{0xa0a4, -0.0000000000000000002778268066994105112144},
-{0xa217, -0.0000000000000000020464316005663896191891},
-{0xa225, -0.0000000000000000022361669807513528951404},
-{0xa22f, -0.0000000000000000023716922523120409493913},
-{0xa241, -0.0000000000000000026156377411212794470430},
-{0xa281, -0.0000000000000000034965520062657517996740},
-{0xa29a, -0.0000000000000000041741783640691920709287},
-{0xa2ce, -0.0000000000000000055836411883003478351384},
-{0xa2d4, -0.0000000000000000057462715141731735002395},
-{0xa2f6, -0.0000000000000000066678433607858522691458},
-{0xa394, -0.0000000000000000160461921527854656233103},
-{0xa3f3, -0.0000000000000000263461127913977577463811},
-{0xa403, -0.0000000000000000284060969191202161709953},
-{0xa416, -0.0000000000000000325260651745651330202236},
-{0xa432, -0.0000000000000000385975973404839578506653},
-{0xa46c, -0.0000000000000000511743425413158092851518},
-{0xa482, -0.0000000000000000563785129692462305683875},
-{0xa4b0, -0.0000000000000000763278329429795121541247},
-{0xa4d4, -0.0000000000000000919403442267707760038320},
-{0xa4dc, -0.0000000000000000954097911787243901926558},
-{0xa52c, -0.0000000000000001491862189340054101194255},
-{0xa595, -0.0000000000000002584737979205442570673767},
-{0xa604, -0.0000000000000004579669976578770729247481},
-{0xa607, -0.0000000000000004683753385137379154912196},
-{0xa62e, -0.0000000000000006036837696399288688553497},
-{0xa641, -0.0000000000000006696032617270475384430028},
-{0xa67b, -0.0000000000000008708311849403571613947861},
-{0xa708, -0.0000000000000018873791418627661187201738},
-{0xa72f, -0.0000000000000024286128663675299321766943},
-{0xa74c, -0.0000000000000028310687127941491780802608},
-{0xa779, -0.0000000000000034555691641457997320685536},
-{0xa7a0, -0.0000000000000044408920985006261616945267},
-{0xa81d, -0.0000000000000087152507433074788423255086},
-{0xa822, -0.0000000000000089928064994637679774314165},
-{0xa83e, -0.0000000000000105471187339389871340245008},
-{0xa8c6, -0.0000000000000219824158875780995003879070},
-{0xa8cd, -0.0000000000000227595720048157090786844492},
-{0xa8fc, -0.0000000000000279776202205539448186755180},
-{0xa909, -0.0000000000000304201108747292892076075077},
-{0xa911, -0.0000000000000321964677141295396722853184},
-{0xa934, -0.0000000000000399680288865056354552507401},
-{0xa9b5, -0.0000000000000803801469828613335266709328},
-{0xa9e9, -0.0000000000001034727858950645895674824715},
-{0xaa3f, -0.0000000000001696420781627239193767309189},
-{0xab0d, -0.0000000000005009326287108706310391426086},
-{0xab40, -0.0000000000006821210263296961784362792969},
-{0xab44, -0.0000000000006963318810448981821537017822},
-{0xab4a, -0.0000000000007176481631177011877298355103},
-{0xab69, -0.0000000000008277822871605167165398597717},
-{0xab6a, -0.0000000000008313350008393172174692153931},
-{0xab85, -0.0000000000009450218385609332472085952759},
-{0xabba, -0.0000000000013216094885137863457202911377},
-{0xac1f, -0.0000000000022595258997171185910701751709},
-{0xac94, -0.0000000000042064129956997931003570556641},
-{0xaca6, -0.0000000000047180037654470652341842651367},
-{0xacbf, -0.0000000000054285465012071654200553894043},
-{0xacf2, -0.0000000000068780536821577697992324829102},
-{0xad7a, -0.0000000000142108547152020037174224853516},
-{0xadaf, -0.0000000000198951966012828052043914794922},
-{0xadfd, -0.0000000000287627699435688555240631103516},
-{0xae3e, -0.0000000000432009983342140913009643554688},
-{0xae45, -0.0000000000447926140623167157173156738281},
-{0xae8a, -0.0000000000627551344223320484161376953125},
-{0xaeda, -0.0000000000991349224932491779327392578125},
-{0xafa2, -0.000000000294676283374428749084472656250},
-{0xaffc, -0.00000000045838532969355583190917968750},
-{0xb016, -0.00000000054569682106375694274902343750},
-{0xb086, -0.0000000009749783203005790710449218750},
-{0xb09d, -0.00000000114232534542679786682128906250},
-{0xb0bd, -0.00000000137515598908066749572753906250},
-{0xb0d1, -0.00000000152067514136433601379394531250},
-{0xb121, -0.0000000023428583517670631408691406250},
-{0xb13a, -0.000000002706656232476234436035156250},
-{0xb259, -0.00000001263106241822242736816406250},
-{0xb25c, -0.000000012805685400962829589843750},
-{0xb2fa, -0.000000029103830456733703613281250},
-{0xb3e2, -0.0000001052394509315490722656250},
-{0xb3ff, -0.00000011874362826347351074218750},
-{0xb44f, -0.0000001927837729454040527343750},
-{0xb4c9, -0.000000374391674995422363281250},
-{0xb4e9, -0.000000433996319770812988281250},
-{0xb4fa, -0.00000046566128730773925781250},
-{0xb5cc, -0.00000151991844177246093750},
-{0xb5df, -0.0000016614794731140136718750},
-{0xb630, -0.00000262260437011718750},
-{0xb648, -0.000002980232238769531250},
-{0xb717, -0.0000090003013610839843750},
-{0xb76f, -0.0000142455101013183593750},
-{0xb78a, -0.00001645088195800781250},
-{0xb7f7, -0.000029444694519042968750},
-{0xb86d, -0.00005650520324707031250},
-{0xb885, -0.0000634193420410156250},
-{0xb8cc, -0.00009727478027343750},
-{0xb8d5, -0.0001015663146972656250},
-{0xb95a, -0.00020790100097656250},
-{0xb967, -0.000220298767089843750},
-{0xb9fc, -0.000480651855468750},
-{0xba87, -0.001029968261718750},
-{0xbac9, -0.001533508300781250},
-{0xbb04, -0.002014160156250},
-{0xbb7b, -0.00382995605468750},
-{0xbbb1, -0.0054016113281250},
-{0xbc7c, -0.0153808593750},
-{0xbc8d, -0.01721191406250},
-{0xbcb7, -0.02233886718750},
-{0xbce7, -0.02819824218750},
-{0xbcfd, -0.03088378906250},
-{0xbd14, -0.03613281250},
-{0xbd44, -0.04785156250},
-{0xbd75, -0.0598144531250},
-{0xbd7e, -0.062011718750},
-{0xbe39, -0.18066406250},
-{0xbebd, -0.3691406250},
-{0xbec2, -0.378906250},
-{0xbee7, -0.4511718750},
-{0xbf16, -0.58593750},
-{0xbf74, -0.9531250},
-{0xbf86, -1.0468750},
-{0xbfab, -1.33593750},
-{0xbfc1, -1.50781250},
-{0xbfc2, -1.5156250},
-{0xbfe1, -1.75781250},
-{0xc01e, -2.468750},
-{0xc02d, -2.7031250},
-{0xc04a, -3.156250},
-{0xc090, -4.50},
-{0xc095, -4.656250},
-{0xc0ce, -6.43750},
-{0xc105, -8.31250},
-{0xc175, -15.31250},
-{0xc18f, -17.8750},
-{0xc1b3, -22.3750},
-{0xc1b6, -22.750},
-{0xc2a0, -80.0},
-{0xc327, -167.0},
-{0xc352, -210.0},
-{0xc380, -256.0},
-{0xc3bb, -374.0},
-{0xc3c4, -392.0},
-{0xc464, -912.0},
-{0xc4ad, -1384.0},
-{0xc4c5, -1576.0},
-{0xc4e7, -1848.0},
-{0xc540, -3072.0},
-{0xc569, -3728.0},
-{0xc56b, -3760.0},
-{0xc57a, -4000.0},
-{0xc5b4, -5760.0},
-{0xc5d0, -6656.0},
-{0xc5d2, -6720.0},
-{0xc5db, -7008.0},
-{0xc5f5, -7840.0},
-{0xc643, -12480.0},
-{0xc655, -13632.0},
-{0xc690, -18432.0},
-{0xc6bc, -24064.0},
-{0xc761, -57600.0},
-{0xc7b4, -92160.0},
-{0xc8a9, -346112.0},
-{0xc8d1, -428032.0},
-{0xc91b, -634880.0},
-{0xca23, -2670592.0},
-{0xca44, -3211264.0},
-{0xcabe, -6225920.0},
-{0xcac8, -6553600.0},
-{0xcb88, -17825792.0},
-{0xcbc4, -25690112.0},
-{0xcbf5, -32112640.0},
-{0xcbfc, -33030144.0},
-{0xcc6e, -62390272.0},
-{0xcc74, -63963136.0},
-{0xcce7, -121110528.0},
-{0xcceb, -123207680.0},
-{0xccee, -124780544.0},
-{0xcd4e, -216006656.0},
-{0xcd5c, -230686720.0},
-{0xcdfd, -530579456.0},
-{0xce03, -549453824.0},
-{0xce35, -759169024.0},
-{0xce6b, -985661440.0},
-{0xce94, -1241513984.0},
-{0xceab, -1434451968.0},
-{0xceb5, -1518338048.0},
-{0xcee0, -1879048192.0},
-{0xcf36, -3053453312.0},
-{0xcf4e, -3456106496.0},
-{0xcfe5, -7683964928.0},
-{0xd023, -10938744832.0},
-{0xd055, -14294188032.0},
-{0xd059, -14562623488.0},
-{0xd0ac, -23085449216.0},
-{0xd0b9, -24830279680.0},
-{0xd0bd, -25367150592.0},
-{0xd14a, -54223962112.0},
-{0xd204, -141733920768.0},
-{0xd216, -161061273600.0},
-{0xd26d, -254476812288.0},
-{0xd37b, -1078036791296.0},
-{0xd37c, -1082331758592.0},
-{0xd471, -4140348473344.0},
-{0xd4a6, -5703716569088.0},
-{0xd4d2, -7215545057280.0},
-{0xd4d3, -7249904795648.0},
-{0xd5a6, -22814866276352.0},
-{0xd5c2, -26663156973568.0},
-{0xd67b, -68994354642944.0},
-{0xd714, -162727720910848.0},
-{0xd754, -233096465088512.0},
-{0xd777, -271579372060672.0},
-{0xd7a5, -362838837166080.0},
-{0xd7bd, -415615395299328.0},
-{0xd809, -602532372021248.0},
-{0xd819, -672901116198912.0},
-{0xd827, -734473767354368.0},
-{0xd82e, -765260092932096.0},
-{0xd859, -954376092909568.0},
-{0xd8a6, -1460151441686528.0},
-{0xd8fc, -2216615441596416.0},
-{0xd93e, -3342515348439040.0},
-{0xd97e, -4468415255281664.0},
-{0xd9a7, -5875790138834944.0},
-{0xd9e0, -7881299347898368.0},
-{0xda46, -13933011347177472.0},
-{0xda4e, -14495961300598784.0},
-{0xda6e, -16747761114284032.0},
-{0xda83, -18436610974547968.0},
-{0xda8e, -19984723346456576.0},
-{0xdb78, -69805794224242688.0},
-{0xdb81, -72620543991349248.0},
-{0xdbaa, -95701492081623040.0},
-{0xdc01, -145241087982698496.0},
-{0xdc2a, -191402984163246080.0},
-{0xdc50, -234187180623265792.0},
-{0xdcf1, -542683755098144768.0},
-{0xdcf9, -560698153607626752.0},
-{0xde68, -4179340454199820288.0},
-{0xdf9c, -22481969339833516032.0},
-{0xdfa6, -23923121220592074752.0},
-{0xdff4, -35164105890508832768.0},
-{0xe018, -43811017175060185088.0},
-{0xe0b2, -102610013910009380864.0},
-{0xe0d9, -125091983249842896896.0},
-{0xe0ff, -146997491837372989440.0},
-{0xe113, -169479461177206505472.0},
-{0xe137, -210984635343052996608.0},
-{0xe1ae, -401216683603182747648.0},
-{0xe261, -1037629354146162278400.0},
-{0xe277, -1139086446551564812288.0},
-{0xe296, -1383505805528216371200.0},
-{0xe2a6, -1531079758117892784128.0},
-{0xe2d0, -1918461383665793368064.0},
-{0xe2dc, -2029141848108050677760.0},
-{0xe2e2, -2084482080329179332608.0},
-{0xe327, -3080606260309495119872.0},
-{0xe336, -3357307421415138394112.0},
-{0xe34d, -3781582535110458081280.0},
-{0xe3e5, -8448608785758974640128.0},
-{0xe457, -15864199903390214389760.0},
-{0xe496, -22136092888451461939200.0},
-{0xe498, -22431240793630814765056.0},
-{0xe51c, -46043073207979040833536.0},
-{0xe53d, -55782954078897684086784.0},
-{0xe547, -58734433130691212345344.0},
-{0xe59a, -90905554795240670363648.0},
-{0xe5ae, -102711471002414783397888.0},
-{0xe5e8, -136948628003219711197184.0},
-{0xe5ed, -139900107055013239455744.0},
-{0xe6f5, -578489894151531538677760.0},
-{0xe730, -831136500985057557610496.0},
-{0xe73c, -887804898779493300174848.0},
-{0xe743, -920861464159580816670720.0},
-{0xe792, -1378931012997936402399232.0},
-{0xe82c, -3248988140214315907022848.0},
-{0xe899, -5780176575032445741563904.0},
-{0xe89a, -5817955506895402903273472.0},
-{0xe8eb, -8878048987794933001748480.0},
-{0xe8ed, -8953606851520847325167616.0},
-{0xe90d, -10653658785353919602098176.0},
-{0xe953, -15942709246167922241437696.0},
-{0xe9a5, -24934095029551726728314880.0},
-{0xea47, -60144059525827801441632256.0},
-{0xeaa2, -97922991388784963151200256.0},
-{0xeae1, -136004154706645782154444800.0},
-{0xeb36, -220024499169862509796524032.0},
-{0xeb58, -261127977036759901736534016.0},
-{0xeb85, -321574268017491360471842816.0},
-{0xec33, -865590886844074489089622016.0},
-{0xec5e, -1073526127817790707139084288.0},
-{0xecc6, -1914938498269572612734582784.0},
-{0xecd2, -2030995376952577013506375680.0},
-{0xed02, -2514565704798428683388846080.0},
-{0xed5f, -4313447324384996895351635968.0},
-{0xed61, -4352132950612665028942233600.0},
-{0xedb2, -6886041468524927779126378496.0},
-{0xedda, -8433466517631653122750283776.0},
-{0xeded, -9168493415957347660971638784.0},
-{0xee2f, -13539969179683846756709171200.0},
-{0xef22, -50136571591057901133414531072.0},
-{0xef24, -50755541610700591270864093184.0},
-{0xef46, -61278031944626323607506649088.0},
-{0xef5a, -67467732141053224982002270208.0},
-{0xef67, -71491037268730710875424423936.0},
-{0xefc6, -122556063889252647215013298176.0},
-{0xf003, -162170145146384816011785273344.0},
-{0xf0a3, -403568452807033969617114497024.0},
-{0xf0b7, -453086054378449180613079465984.0},
-{0xf0e1, -557073017678421123704605900800.0},
-{0xf14d, -1015110832214011825417281863680.0},
-{0xf18c, -1386492843999625907887019130880.0},
-{0xf1c1, -1911379420656627144444247801856.0},
-{0xf21d, -3109705378684875250546600050688.0},
-{0xf22a, -3367196906856234347725617889280.0},
-{0xf22c, -3406810988113366516522389864448.0},
-{0xf239, -3664302516284725613701407703040.0},
-{0xf240, -3802951800684688204490109616128.0},
-{0xf267, -4575426385198765496027163131904.0},
-{0xf270, -4753689750855860255612637020160.0},
-{0xf27f, -5050795360284351521588426833920.0},
-{0xf2df, -8833940120340473641680150462464.0},
-{0xf2f8, -9824292151768777861599449841664.0},
-{0xf335, -14340297415081845104431455010816.0},
-{0xf368, -18380933703309326321702196477952.0},
-{0xf3ba, -29472876455306333584798349524992.0},
-{0xf3c1, -30582070730506034311107964829696.0},
-{0xf418, -48170722808672717256874721804288.0},
-{0xf435, -57361189660327380417725820043264.0},
-{0xf494, -93806144416888975710756037197824.0},
-{0xf49b, -98242921517687778615994498416640.0},
-{0xf4a0, -101412048018258352119736256430080.0},
-{0xf4ec, -149582770826931069376610978234368.0},
-{0xf50a, -174935782831495657406545042341888.0},
-{0xf563, -287756686251808074139751627620352.0},
-{0xf583, -332124457259796103192136239808512.0},
-{0xf5a0, -405648192073033408478945025720320.0},
-{0xf5f0, -608472288109550112718417538580480.0},
-{0xf60c, -709884336127808464838153795010560.0},
-{0xf614, -750449155335111805686048297582592.0},
-{0xf6d9, -2200641441996206240998276764532736.0},
-{0xf6f2, -2454171562041852121297617405607936.0},
-{0xf700, -2596148429267413814265248164610048.0},
-{0xf730, -3569704090242693994614716226338816.0},
-{0xf77b, -5090884810516569276410760072790016.0},
-{0xf7a0, -6490371073168534535663120411525120.0},
-{0xf7ae, -7058278542070781307533643447533568.0},
-{0xf7da, -8843130587192128304841001560702976.0},
-{0xf7f2, -9816686248167408485190469622431744.0},
-{0xf7f3, -9857251067374711826038364125003776.0},
-{0xf7fa, -10141204801825835211973625643008000.0},
-{0xf831, -14359945999385382660154653910499328.0},
-{0xf84d, -16631575874994369747636746054533120.0},
-{0xf85c, -17848520451213469973073581131694080.0},
-{0xf8ac, -27908595614624698503351417769558016.0},
-{0xf90e, -46081634619496595203208154921828352.0},
-{0xf93f, -61983043748759504815582799930064896.0},
-{0xf97c, -81778675521923535149355317185216512.0},
-{0xf98e, -92163269238993190406416309843656704.0},
-{0xf9dc, -142788163609707759784588649053552640.0},
-{0xf9e7, -149927571790193147773818081506230272.0},
-{0xfa00, -166153499473114484112975882535043072.0},
-{0xfa45, -255720620282840260705126944214089728.0},
-{0xfa78, -321922405229159312968890772411645952.0},
-{0xfae3, -589325693443702935838211333366480896.0},
-{0xfb2b, -887882762809455524478714872296636416.0},
-{0xfb59, -1126728418302057595391117703440760832.0},
-{0xfb82, -1349997183219055183417929045597224960.0},
-{0xfb94, -1536919870126308978045026913449148416.0},
-{0xfb96, -1557689057560448288559148898766028800.0},
-{0xfb9e, -1640765807297005530615636840033550336.0},
-{0xfbad, -1796534713053050359471551729910153216.0},
-{0xfbc2, -2014611181111513119869832575737397248.0},
-{0xfbde, -2305379805189463467067540370173722624.0},
-{0xfbfd, -2627302210418622780036431142585368576.0},
-{0xfc37, -3800761300447493824084323312989110272.0},
-{0xfd39, -15369198701263089780450269134491484160.0},
-{0xfd4a, -16781503446784562895410564136039350272.0},
-{0xfd84, -21932261930451111902912816494625685504.0},
-{0xfdcc, -33895313892515354759047080037148786688.0},
-{0xff1b, -206030339346661960300090094343453409280.0},
-{0xff45, -261857915169628426962049990875227881472.0},
-{0xff51, -277808651119047417436895675598592016384.0}
+static const std::map<uint16_t, float> bf16_lut = {
+    {0x0000, 0.0},
+    {0x002d, 0.00000000000000000000000000000000000000413259732711},
+    {0x004e, 0.00000000000000000000000000000000000000716316870032},
+    {0x005b, 0.00000000000000000000000000000000000000835703015038},
+    {0x00cd, 0.00000000000000000000000000000000000001882627671239},
+    {0x00ce, 0.00000000000000000000000000000000000001891811220855},
+    {0x0170, 0.00000000000000000000000000000000000004408103815584},
+    {0x01be, 0.00000000000000000000000000000000000006979497708007},
+    {0x01fe, 0.00000000000000000000000000000000000009330486409652},
+    {0x0211, 0.00000000000000000000000000000000000010652917554327},
+    {0x028f, 0.00000000000000000000000000000000000021011961520948},
+    {0x02bc, 0.00000000000000000000000000000000000027624117244324},
+    {0x02e8, 0.00000000000000000000000000000000000034089336173846},
+    {0x039c, 0.00000000000000000000000000000000000091688559364138},
+    {0x03a9, 0.00000000000000000000000000000000000099329272644483},
+    {0x03cf, 0.00000000000000000000000000000000000121663665310107},
+    {0x03da, 0.00000000000000000000000000000000000128128884239629},
+    {0x03f3, 0.00000000000000000000000000000000000142822563624908},
+    {0x0427, 0.00000000000000000000000000000000000196307556587322},
+    {0x044c, 0.00000000000000000000000000000000000239800847567747},
+    {0x0485, 0.00000000000000000000000000000000000312681497318728},
+    {0x0498, 0.00000000000000000000000000000000000357350282649975},
+    {0x04da, 0.00000000000000000000000000000000000512515536958517},
+    {0x051b, 0.00000000000000000000000000000000000728806497509818},
+    {0x0533, 0.00000000000000000000000000000000000841653955188758},
+    {0x0536, 0.00000000000000000000000000000000000855759887398625},
+    {0x0577, 0.00000000000000000000000000000000001161388418612420},
+    {0x0587, 0.00000000000000000000000000000000001269533898888071},
+    {0x065a, 0.00000000000000000000000000000000004100124295668139},
+    {0x0735, 0.00000000000000000000000000000000013616926559925378},
+    {0x075e, 0.00000000000000000000000000000000016701423736483061},
+    {0x0765, 0.00000000000000000000000000000000017228045205651446},
+    {0x076b, 0.00000000000000000000000000000000017679435036367204},
+    {0x07ba, 0.00000000000000000000000000000000027986169504377021},
+    {0x07d1, 0.00000000000000000000000000000000031446824873197835},
+    {0x07db, 0.00000000000000000000000000000000032951457642250363},
+    {0x07eb, 0.00000000000000000000000000000000035358870072734408},
+    {0x080d, 0.00000000000000000000000000000000042430644087281290},
+    {0x0841, 0.00000000000000000000000000000000058078824885427581},
+    {0x08ff, 0.00000000000000000000000000000000153472542443357857},
+    {0x09e1, 0.00000000000000000000000000000000541667796858910084},
+    {0x0a0b, 0.00000000000000000000000000000000669260655674564459},
+    {0x0a0f, 0.00000000000000000000000000000000688519955118436817},
+    {0x0a12, 0.00000000000000000000000000000000702964429701341086},
+    {0x0a1d, 0.00000000000000000000000000000000755927503171990072},
+    {0x0aa2, 0.00000000000000000000000000000001560003254953661041},
+    {0x0aaf, 0.00000000000000000000000000000001685188701338831371},
+    {0x0ab4, 0.00000000000000000000000000000001733336949948512268},
+    {0x0ade, 0.00000000000000000000000000000002137782238269831797},
+    {0x0b6e, 0.00000000000000000000000000000004583713267641621330},
+    {0x0bb6, 0.00000000000000000000000000000007010384997569538505},
+    {0x0c8c, 0.00000000000000000000000000000021570415377137041554},
+    {0x0cf7, 0.00000000000000000000000000000038056375701091780456},
+    {0x0d06, 0.00000000000000000000000000000041291938007662336690},
+    {0x0d20, 0.00000000000000000000000000000049303806576313237838},
+    {0x0d5e, 0.00000000000000000000000000000068409031624634617501},
+    {0x0d60, 0.00000000000000000000000000000069025329206838532974},
+    {0x0d6a, 0.00000000000000000000000000000072106817117858110338},
+    {0x0d76, 0.00000000000000000000000000000075804602611081603176},
+    {0x0d7e, 0.00000000000000000000000000000078269792939897265068},
+    {0x0dcc, 0.00000000000000000000000000000125724706769598756487},
+    {0x0dfc, 0.00000000000000000000000000000155306990715386699190},
+    {0x0e09, 0.00000000000000000000000000000168865537523872839596},
+    {0x0e3c, 0.00000000000000000000000000000231727890908672217840},
+    {0x0e69, 0.00000000000000000000000000000287194673307024610408},
+    {0x0f42, 0.00000000000000000000000000000956493847580476814062},
+    {0x0f8f, 0.00000000000000000000000000001410088868082558602173},
+    {0x0fa2, 0.00000000000000000000000000001597443333072548905959},
+    {0x1007, 0.00000000000000000000000000002662405555120914843265},
+    {0x1015, 0.00000000000000000000000000002938506871948268975159},
+    {0x1030, 0.00000000000000000000000000003470987982972451943812},
+    {0x1042, 0.00000000000000000000000000003825975390321907256247},
+    {0x1056, 0.00000000000000000000000000004220405842932413158953},
+    {0x1166, 0.00000000000000000000000000018143800820083271524470},
+    {0x1182, 0.00000000000000000000000000020510383535746306940705},
+    {0x128a, 0.00000000000000000000000000087090243936399703317455},
+    {0x1295, 0.00000000000000000000000000094032219902344607205078},
+    {0x129e, 0.00000000000000000000000000099712018419935892204042},
+    {0x12b0, 0.00000000000000000000000000111071615455118462201971},
+    {0x12bf, 0.00000000000000000000000000120537946317770603866912},
+    {0x1315, 0.00000000000000000000000000188064439804689214410156},
+    {0x1343, 0.00000000000000000000000000246124602428955683288459},
+    {0x13e2, 0.00000000000000000000000000570504206655835737673762},
+    {0x13ee, 0.00000000000000000000000000600796465416322591001572},
+    {0x1445, 0.00000000000000000000000000994595829302651684263107},
+    {0x1469, 0.00000000000000000000000001176349381865572804229970},
+    {0x14bd, 0.00000000000000000000000001908412301910671759652054},
+    {0x1541, 0.00000000000000000000000003897603960515975128178268},
+    {0x1546, 0.00000000000000000000000003998578156384264639270970},
+    {0x1569, 0.00000000000000000000000004705397527462291216919879},
+    {0x16a1, 0.00000000000000000000000026010952855671378057479844},
+    {0x16b6, 0.00000000000000000000000029403685836845905630194606},
+    {0x16f4, 0.00000000000000000000000039420326066980225130590570},
+    {0x1703, 0.00000000000000000000000042328382907986963050060367},
+    {0x1720, 0.00000000000000000000000051698788284564229679463043},
+    {0x178e, 0.00000000000000000000000091765349205101507681046902},
+    {0x17f6, 0.00000000000000000000000158973773975035006264348858},
+    {0x18e8, 0.00000000000000000000000599705944100945064281771302},
+    {0x18f1, 0.00000000000000000000000622970398828998967637529671},
+    {0x1927, 0.00000000000000000000000863369764352222635647032822},
+    {0x1a39, 0.00000000000000000000003825710333057752996280265201},
+    {0x1a60, 0.00000000000000000000004632211430296954979279888676},
+    {0x1a69, 0.00000000000000000000004818327068121386206125955631},
+    {0x1b1f, 0.00000000000000000000013152171739593140030455398204},
+    {0x1b43, 0.00000000000000000000016130021944784039659992469495},
+    {0x1ba0, 0.00000000000000000000026469779601696885595885078146},
+    {0x1bc0, 0.00000000000000000000031763735522036262715062093775},
+    {0x1bea, 0.00000000000000000000038712052667481695183981926789},
+    {0x1c25, 0.00000000000000000000054593920428499826541512973677},
+    {0x1c32, 0.00000000000000000000058895259613775570450844298875},
+    {0x1cbe, 0.00000000000000000000125731453108060206580454121195},
+    {0x1ccd, 0.00000000000000000000135657620458696538678911025499},
+    {0x1cd6, 0.00000000000000000000141613320869078337937985168082},
+    {0x1d23, 0.00000000000000000000215728703753829617606463386892},
+    {0x1d38, 0.00000000000000000000243521972335611347482142718945},
+    {0x1dce, 0.00000000000000000000545277459794955843275232609813},
+    {0x1e7c, 0.00000000000000000001334076891925523034032607938570},
+    {0x1e9e, 0.00000000000000000001672890070827243169659936938842},
+    {0x1eb1, 0.00000000000000000001874060395800139500188663532754},
+    {0x1f33, 0.00000000000000000003790472438962994017330743190541},
+    {0x1f56, 0.00000000000000000004531626267810506814015525378636},
+    {0x1fa8, 0.00000000000000000007115076756936122848173909005709},
+    {0x1fed, 0.00000000000000000010037340424963459017959621633054},
+    {0x2001, 0.00000000000000000010926725019580474373981360258767},
+    {0x213f, 0.00000000000000000064713317170228545904819839051925},
+    {0x2154, 0.00000000000000000071828393927164668752993748057634},
+    {0x216b, 0.00000000000000000079621097041904231872422315063886},
+    {0x219f, 0.00000000000000000107742590890747003129490622086450},
+    {0x2225, 0.00000000000000000223616698075135289514037140179425},
+    {0x2229, 0.00000000000000000229037708937562811684074404183775},
+    {0x2274, 0.00000000000000000330681662608078852372273104265332},
+    {0x227b, 0.00000000000000000340168431617327016169838316272944},
+    {0x2294, 0.00000000000000000401154803819636640582757536321878},
+    {0x22b7, 0.00000000000000000496022493912118278558409656397998},
+    {0x2379, 0.00000000000000001349831704744453020339278737083077},
+    {0x2389, 0.00000000000000001485356976305141074590210337191820},
+    {0x2464, 0.00000000000000004943961906533900219073984771966934},
+    {0x2475, 0.00000000000000005312590645178971726636518724262714},
+    {0x2491, 0.00000000000000006288372600415925717243226245045662},
+    {0x24f6, 0.00000000000000010668549377257363630633335560560226},
+    {0x2567, 0.00000000000000020036056147532121940457727760076523},
+    {0x256a, 0.00000000000000020296264668928643004619516432285309},
+    {0x257f, 0.00000000000000022117724318704290453752037137746811},
+    {0x25f9, 0.00000000000000043194614551822496650856919586658478},
+    {0x263b, 0.00000000000000064878658001532585331005975604057312},
+    {0x270c, 0.00000000000000194289029309402394574135541915893555},
+    {0x272c, 0.00000000000000238697950294408656191080808639526367},
+    {0x2745, 0.00000000000000273392419813944798079319298267364502},
+    {0x2791, 0.00000000000000402455846426619245903566479682922363},
+    {0x2792, 0.00000000000000405231403988182137254625558853149414},
+    {0x27a0, 0.00000000000000444089209850062616169452667236328125},
+    {0x27a3, 0.00000000000000452415882534751290222629904747009277},
+    {0x27d4, 0.00000000000000588418203051332966424524784088134766},
+    {0x27dd, 0.00000000000000613398221105398988584056496620178223},
+    {0x2821, 0.00000000000000893729534823251015041023492813110352},
+    {0x28d9, 0.00000000000002409183963436589692719280719757080078},
+    {0x2981, 0.00000000000005728750807065807748585939407348632812},
+    {0x29ca, 0.00000000000008970602038971264846622943878173828125},
+    {0x2a5b, 0.00000000000019451107391432742588222026824951171875},
+    {0x2aa8, 0.00000000000029842794901924207806587219238281250},
+    {0x2ac4, 0.000000000000348165940522449091076850891113281250},
+    {0x2ae7, 0.00000000000041033842990145785734057426452636718750},
+    {0x2af1, 0.00000000000042810199829546036198735237121582031250},
+    {0x2afe, 0.0000000000004511946372076636180281639099121093750},
+    {0x2b24, 0.00000000000058264504332328215241432189941406250},
+    {0x2b4c, 0.00000000000072475359047530218958854675292968750},
+    {0x2b85, 0.000000000000945021838560933247208595275878906250},
+    {0x2bed, 0.000000000001683986283751437440514564514160156250},
+    {0x2c18, 0.00000000000216004991671070456504821777343750},
+    {0x2cdf, 0.0000000000063380412029800936579704284667968750},
+    {0x2d6b, 0.000000000013358203432289883494377136230468750},
+    {0x2d96, 0.0000000000170530256582424044609069824218750},
+    {0x2da2, 0.0000000000184172677109017968177795410156250},
+    {0x2db8, 0.00000000002091837814077734947204589843750},
+    {0x2de1, 0.00000000002557953848736360669136047363281250},
+    {0x2e90, 0.00000000006548361852765083312988281250},
+    {0x2ea3, 0.000000000074123818194493651390075683593750},
+    {0x2ef0, 0.00000000010913936421275138854980468750},
+    {0x2f09, 0.00000000012460077414289116859436035156250},
+    {0x2f6b, 0.00000000021373125491663813591003417968750},
+    {0x303b, 0.000000000680302036926150321960449218750},
+    {0x308f, 0.00000000104046193882822990417480468750},
+    {0x309c, 0.000000001135049387812614440917968750},
+    {0x30a9, 0.00000000122963683679699897766113281250},
+    {0x312a, 0.000000002473825588822364807128906250},
+    {0x313d, 0.0000000027503119781613349914550781250},
+    {0x3159, 0.0000000031577656045556068420410156250},
+    {0x31c6, 0.00000000576255843043327331542968750},
+    {0x3212, 0.0000000084983184933662414550781250},
+    {0x3245, 0.00000001146690919995307922363281250},
+    {0x329b, 0.0000000180443748831748962402343750},
+    {0x32ba, 0.000000021653249859809875488281250},
+    {0x32cc, 0.00000002374872565269470214843750},
+    {0x3332, 0.00000004144385457038879394531250},
+    {0x33c4, 0.000000091269612312316894531250},
+    {0x3424, 0.00000015273690223693847656250},
+    {0x3589, 0.0000010207295417785644531250},
+    {0x3594, 0.00000110268592834472656250},
+    {0x368b, 0.00000414252281188964843750},
+    {0x36a0, 0.000004768371582031250},
+    {0x36e9, 0.00000694394111633300781250},
+    {0x36ed, 0.00000706315040588378906250},
+    {0x3750, 0.000012397766113281250},
+    {0x375f, 0.0000132918357849121093750},
+    {0x37ce, 0.00002455711364746093750},
+    {0x37d2, 0.00002503395080566406250},
+    {0x37f0, 0.00002861022949218750},
+    {0x380e, 0.0000338554382324218750},
+    {0x3826, 0.0000395774841308593750},
+    {0x387f, 0.00006079673767089843750},
+    {0x38e1, 0.0001072883605957031250},
+    {0x38e9, 0.0001111030578613281250},
+    {0x3964, 0.0002174377441406250},
+    {0x3994, 0.000282287597656250},
+    {0x3a26, 0.000633239746093750},
+    {0x3a2c, 0.00065612792968750},
+    {0x3a6a, 0.000892639160156250},
+    {0x3a85, 0.001014709472656250},
+    {0x3ab9, 0.001411437988281250},
+    {0x3aba, 0.00141906738281250},
+    {0x3af7, 0.001884460449218750},
+    {0x3b03, 0.00199890136718750},
+    {0x3bb3, 0.0054626464843750},
+    {0x3bbf, 0.0058288574218750},
+    {0x3be8, 0.0070800781250},
+    {0x3c06, 0.00817871093750},
+    {0x3c29, 0.010314941406250},
+    {0x3c3f, 0.011657714843750},
+    {0x3c73, 0.014831542968750},
+    {0x3ce9, 0.02844238281250},
+    {0x3cfa, 0.0305175781250},
+    {0x3cfb, 0.03063964843750},
+    {0x3d0f, 0.0349121093750},
+    {0x3d2a, 0.041503906250},
+    {0x3d43, 0.0476074218750},
+    {0x3dd0, 0.10156250},
+    {0x3dd9, 0.105957031250},
+    {0x3de9, 0.113769531250},
+    {0x3df9, 0.121582031250},
+    {0x3e1e, 0.1542968750},
+    {0x3e77, 0.24121093750},
+    {0x3e95, 0.2910156250},
+    {0x3f38, 0.718750},
+    {0x3fb3, 1.39843750},
+    {0x3fc5, 1.53906250},
+    {0x3fd3, 1.64843750},
+    {0x3fd7, 1.67968750},
+    {0x400b, 2.1718750},
+    {0x40bf, 5.968750},
+    {0x40c7, 6.218750},
+    {0x4123, 10.18750},
+    {0x412b, 10.68750},
+    {0x41bf, 23.8750},
+    {0x41ca, 25.250},
+    {0x421b, 38.750},
+    {0x4226, 41.50},
+    {0x42a7, 83.50},
+    {0x42b7, 91.50},
+    {0x4311, 145.0},
+    {0x431f, 159.0},
+    {0x4334, 180.0},
+    {0x434f, 207.0},
+    {0x43b1, 354.0},
+    {0x43e5, 458.0},
+    {0x4476, 984.0},
+    {0x4496, 1200.0},
+    {0x44a4, 1312.0},
+    {0x458b, 4448.0},
+    {0x45a9, 5408.0},
+    {0x45df, 7136.0},
+    {0x45f6, 7872.0},
+    {0x45fa, 8000.0},
+    {0x4602, 8320.0},
+    {0x4640, 12288.0},
+    {0x4648, 12800.0},
+    {0x46a7, 21376.0},
+    {0x46b1, 22656.0},
+    {0x4742, 49664.0},
+    {0x4744, 50176.0},
+    {0x475e, 56832.0},
+    {0x477a, 64000.0},
+    {0x4837, 187392.0},
+    {0x488a, 282624.0},
+    {0x488f, 292864.0},
+    {0x48ea, 479232.0},
+    {0x495c, 901120.0},
+    {0x49aa, 1392640.0},
+    {0x49b9, 1515520.0},
+    {0x4a1e, 2588672.0},
+    {0x4a2b, 2801664.0},
+    {0x4a4c, 3342336.0},
+    {0x4ab6, 5963776.0},
+    {0x4b34, 11796480.0},
+    {0x4b73, 15925248.0},
+    {0x4b7b, 16449536.0},
+    {0x4bcd, 26869760.0},
+    {0x4bd0, 27262976.0},
+    {0x4c07, 35389440.0},
+    {0x4c17, 39583744.0},
+    {0x4c53, 55312384.0},
+    {0x4cad, 90701824.0},
+    {0x4d1c, 163577856.0},
+    {0x4dc0, 402653184.0},
+    {0x4dde, 465567744.0},
+    {0x4eef, 2004877312.0},
+    {0x4efc, 2113929216.0},
+    {0x4f12, 2449473536.0},
+    {0x4f2f, 2936012800.0},
+    {0x4f92, 4898947072.0},
+    {0x4fad, 5804916736.0},
+    {0x4fdc, 7381975040.0},
+    {0x4feb, 7885291520.0},
+    {0x5076, 16508780544.0},
+    {0x5083, 17582522368.0},
+    {0x5215, 159987531776.0},
+    {0x52a9, 362924736512.0},
+    {0x5394, 1271310319616.0},
+    {0x53a0, 1374389534720.0},
+    {0x53b7, 1571958030336.0},
+    {0x540e, 2439541424128.0},
+    {0x542f, 3006477107200.0},
+    {0x5465, 3934190043136.0},
+    {0x5529, 11613591568384.0},
+    {0x554c, 14018773254144.0},
+    {0x5596, 20615843020800.0},
+    {0x55ae, 23914377904128.0},
+    {0x55be, 26113401159680.0},
+    {0x55da, 29961691856896.0},
+    {0x568d, 77515569758208.0},
+    {0x5690, 79164837199872.0},
+    {0x56ad, 95107755802624.0},
+    {0x5718, 167125767421952.0},
+    {0x571c, 171523813933056.0},
+    {0x571d, 172623325560832.0},
+    {0x5826, 730075720843264.0},
+    {0x587c, 1108307720798208.0},
+    {0x5890, 1266637395197952.0},
+    {0x58a8, 1477743627730944.0},
+    {0x58f6, 2163838883463168.0},
+    {0x5966, 4046202790215680.0},
+    {0x5985, 4679521487814656.0},
+    {0x59ad, 6086896371367936.0},
+    {0x59b0, 6192449487634432.0},
+    {0x59bc, 6614661952700416.0},
+    {0x5a93, 20688410788233216.0},
+    {0x5ab0, 24769797950537728.0},
+    {0x5ab6, 25614222880669696.0},
+    {0x5ae3, 31947409856659456.0},
+    {0x5af5, 34480684647055360.0},
+    {0x5afd, 35606584553897984.0},
+    {0x5bb6, 102456891522678784.0},
+    {0x5c3d, 212795082393255936.0},
+    {0x5d57, 968273919884656640.0},
+    {0x5d76, 1107885508333142016.0},
+    {0x5d86, 1206964700135292928.0},
+    {0x5db2, 1603281467343896576.0},
+    {0x5dc1, 1738389456165011456.0},
+    {0x5dc5, 1774418253183975424.0},
+    {0x5e5d, 3981182070595518464.0},
+    {0x5fa4, 23634890844440363008.0},
+    {0x5fa8, 24211351596743786496.0},
+    {0x5ff8, 35740566642812256256.0},
+    {0x6006, 38622870404329373696.0},
+    {0x6051, 60240148615707754496.0},
+    {0x60ed, 136621198295911366656.0},
+    {0x610b, 160256089140351729664.0},
+    {0x6114, 170632382681813352448.0},
+    {0x613b, 215596321361480384512.0},
+    {0x6148, 230584300921369395200.0},
+    {0x61ad, 398910840593969053696.0},
+    {0x61fd, 583378281331064569856.0},
+    {0x629f, 1466516153859909353472.0},
+    {0x62a4, 1512633014044183232512.0},
+    {0x62fb, 2315066381250548727808.0},
+    {0x634b, 3744689046963038978048.0},
+    {0x635a, 4021390208068682252288.0},
+    {0x635f, 4113623928437230010368.0},
+    {0x637c, 4648579506574807007232.0},
+    {0x638d, 5201981828786093555712.0},
+    {0x63a9, 6234999496913828446208.0},
+    {0x6469, 17192365476697302106112.0},
+    {0x64c0, 28334198897217871282176.0},
+    {0x64d1, 30842956091242370301952.0},
+    {0x64dd, 32613843522318487257088.0},
+    {0x64ee, 35122600716342986276864.0},
+    {0x64ef, 35270174668932662689792.0},
+    {0x64fb, 37041062100008779644928.0},
+    {0x6510, 42501298345826806923264.0},
+    {0x6581, 76148159536273029070848.0},
+    {0x65d4, 125142711796045598162944.0},
+    {0x6612, 172366376624742050299904.0},
+    {0x661c, 184172292831916163334144.0},
+    {0x66c6, 467514281804094876155904.0},
+    {0x66ed, 559600428220052957822976.0},
+    {0x66f9, 587934627117270829105152.0},
+    {0x6703, 618630009255923522994176.0},
+    {0x6752, 991696961402625494876160.0},
+    {0x6797, 1426154677826632854536192.0},
+    {0x679c, 1473378342655329306673152.0},
+    {0x67e3, 2143954383222818927017984.0},
+    {0x6862, 4269019300514159273181184.0},
+    {0x692f, 13222626152035006598348800.0},
+    {0x693d, 14280436244197807126216704.0},
+    {0x6943, 14733783426553293066731520.0},
+    {0x695e, 16773845747152979799048192.0},
+    {0x69ec, 35663311678631560653832192.0},
+    {0x69f4, 36872237498246189828538368.0},
+    {0x6a5c, 66490920078804604608839680.0},
+    {0x6a7a, 75557863725914323419136000.0},
+    {0x6ac3, 117870267412426344533852160.0},
+    {0x6ad8, 130563988518379950868267008.0},
+    {0x6ae3, 137213080526260411329150976.0},
+    {0x6b37, 221233424989477138971230208.0},
+    {0x6b77, 298604677444813406152425472.0},
+    {0x6bd7, 519838102434290545123655680.0},
+    {0x6be7, 558523728661958678714253312.0},
+    {0x6c15, 720519788490318988124880896.0},
+    {0x6c33, 865590886844074489089622016.0},
+    {0x6c43, 942962139299410756270817280.0},
+    {0x6c45, 952633545856327789668466688.0},
+    {0x6c48, 967140655691703339764940800.0},
+    {0x6da6, 6421813953792910176039206912.0},
+    {0x6dba, 7195526478346272847851159552.0},
+    {0x6def, 9245864668412683928152834048.0},
+    {0x6e24, 12688885402675147817716023296.0},
+    {0x6e28, 12998370412496492886440804352.0},
+    {0x6e3d, 14623166714058554497245904896.0},
+    {0x6ea0, 24758800785707605497982484480.0},
+    {0x6eef, 36983458673650735712611336192.0},
+    {0x6ef9, 38530883722757461056235241472.0},
+    {0x6f55, 65920307091946499638378364928.0},
+    {0x6f5c, 68086702160695915119451832320.0},
+    {0x6f65, 70872067249088020737974861824.0},
+    {0x6f9e, 97797263103545041717030813696.0},
+    {0x6fdc, 136173404321391830238903664640.0},
+    {0x70ab, 423375493435600054015500484608.0},
+    {0x70dc, 544693617285567320955614658560.0},
+    {0x714a, 1000255551742587262118492372992.0},
+    {0x71ba, 1842054778456645849049896845312.0},
+    {0x71d3, 2089642786313721904029721690112.0},
+    {0x71ee, 2357037834799364043407932522496.0},
+    {0x71f6, 2436265997313628381001476472832.0},
+    {0x7251, 4139671491370311639262671405056.0},
+    {0x72b8, 7288990951312319058606043430912.0},
+    {0x731f, 12597277839768029677373488103424.0},
+    {0x7328, 13310331302396408715715383656448.0},
+    {0x7356, 16954826778052568245018405371904.0},
+    {0x7358, 17113283103081096920205493272576.0},
+    {0x7375, 19410899815994762710418267832320.0},
+    {0x737c, 19965496953594613073573075484672.0},
+    {0x7467, 73206822163180247936434610110464.0},
+    {0x74a4, 103947349218714810922729662840832.0},
+    {0x74b1, 112187078120198302032458233675776.0},
+    {0x7530, 223106505640168374663419764146176.0},
+    {0x7563, 287756686251808074139751627620352.0},
+    {0x756f, 302968493454546826957712066084864.0},
+    {0x7571, 305503794655003285760705472495616.0},
+    {0x7626, 841719998551544322593810928369664.0},
+    {0x7660, 1135814937804493543741046072016896.0},
+    {0x7675, 1242297588223664813466769141268480.0},
+    {0x76b2, 1805134454724998667731305364455424.0},
+    {0x772c, 3488574451828087312918927221194752.0},
+    {0x77b2, 7220537818899994670925221457821696.0},
+    {0x783a, 15090112745116842795416754956795904.0},
+    {0x795d, 71718600358512306619077480547352576.0},
+    {0x796e, 77235415770705560974391132897148928.0},
+    {0x798c, 90865195024359483499283685761351680.0},
+    {0x79af, 113581493780449354374104607201689600.0},
+    {0x7aa5, 428364490829123279353765947160657920.0},
+    {0x7acf, 537402724858354659552906370074279936.0},
+    {0x7adc, 571152654438831039138354596214210560.0},
+    {0x7b07, 700960075902201729851617004444712960.0},
+    {0x7b0f, 742498450770480350879860975078473728.0},
+    {0x7b12, 758075341346084833765452464066134016.0},
+    {0x7b30, 913844247102129662621367353942736896.0},
+    {0x7bc2, 2014611181111513119869832575737397248.0},
+    {0x7bd1, 2170380086867557948725747465614000128.0},
+    {0x7bd5, 2211918461735836569753991436247760896.0},
+    {0x7cf1, 10010748343255147667806796922736345088.0},
+    {0x7d2f, 14538431203897517359885389721816268800.0},
+    {0x7da0, 26584559915698317458076141205606891520.0},
+    {0x7e58, 71778311772385457136805581255138607104.0},
+    {0x7e81, 85735205728127073802295555388082225152.0},
+    {0x7f09, 182104235422533474587821567258407206912.0},
+    {0x7f24, 217993391308726203156224357885976510464.0},
+    {0x7f86, std::numeric_limits<float>::quiet_NaN()},
+    {0x7f88, std::numeric_limits<float>::quiet_NaN()},
+    {0x7f8f, std::numeric_limits<float>::quiet_NaN()},
+    {0x7fa0, std::numeric_limits<float>::quiet_NaN()},
+    {0x7fcd, std::numeric_limits<float>::quiet_NaN()},
+    {0x8023, -0.00000000000000000000000000000000000000321424236553},
+    {0x8074, -0.00000000000000000000000000000000000001065291755433},
+    {0x8080, -0.00000000000000000000000000000000000001175494350822},
+    {0x80a5, -0.00000000000000000000000000000000000001515285686607},
+    {0x80d2, -0.00000000000000000000000000000000000001928545419318},
+    {0x80fd, -0.00000000000000000000000000000000000002323438052797},
+    {0x810a, -0.00000000000000000000000000000000000002534659693961},
+    {0x8124, -0.00000000000000000000000000000000000003012204273982},
+    {0x81e3, -0.00000000000000000000000000000000000008338663051146},
+    {0x81f1, -0.00000000000000000000000000000000000008852941829630},
+    {0x8285, -0.00000000000000000000000000000000000019542593582421},
+    {0x828b, -0.00000000000000000000000000000000000020424214345537},
+    {0x829f, -0.00000000000000000000000000000000000023362950222593},
+    {0x82bc, -0.00000000000000000000000000000000000027624117244324},
+    {0x82ee, -0.00000000000000000000000000000000000034970956936963},
+    {0x82f9, -0.00000000000000000000000000000000000036587261669344},
+    {0x8393, -0.00000000000000000000000000000000000086398834785438},
+    {0x8394, -0.00000000000000000000000000000000000086986581960849},
+    {0x843e, -0.00000000000000000000000000000000000223343926656235},
+    {0x8451, -0.00000000000000000000000000000000000245678319321858},
+    {0x847f, -0.00000000000000000000000000000000000299751059459683},
+    {0x8483, -0.00000000000000000000000000000000000307979519915439},
+    {0x84a0, -0.00000000000000000000000000000000000376158192263132},
+    {0x84aa, -0.00000000000000000000000000000000000399668079279578},
+    {0x84ba, -0.00000000000000000000000000000000000437283898505891},
+    {0x84e4, -0.00000000000000000000000000000000000536025423974963},
+    {0x8510, -0.00000000000000000000000000000000000677084746073638},
+    {0x854d, -0.00000000000000000000000000000000000963905367674276},
+    {0x8557, -0.00000000000000000000000000000000001010925141707167},
+    {0x8584, -0.00000000000000000000000000000000001241322034468336},
+    {0x85b7, -0.00000000000000000000000000000000001720923729603829},
+    {0x8656, -0.00000000000000000000000000000000004024892657215512},
+    {0x87c7, -0.00000000000000000000000000000000029942192104145307},
+    {0x88d9, -0.00000000000000000000000000000000130602124353759431},
+    {0x896c, -0.00000000000000000000000000000000284074666797117288},
+    {0x89a0, -0.00000000000000000000000000000000385185988877447171},
+    {0x89b2, -0.00000000000000000000000000000000428519412626159977},
+    {0x89fe, -0.00000000000000000000000000000000611482757342947383},
+    {0x8a31, -0.00000000000000000000000000000000852224000391351865},
+    {0x8a55, -0.00000000000000000000000000000001025557695386203092},
+    {0x8a68, -0.00000000000000000000000000000001117039367744596795},
+    {0x8b4a, -0.00000000000000000000000000000003890378487662216423},
+    {0x8b4c, -0.00000000000000000000000000000003928897086549961140},
+    {0x8b5b, -0.00000000000000000000000000000004217786578208046518},
+    {0x8bbf, -0.00000000000000000000000000000007357052387559240959},
+    {0x8bff, -0.00000000000000000000000000000009822242716374902851},
+    {0x8c43, -0.00000000000000000000000000000015022253566220439654},
+    {0x8c6b, -0.00000000000000000000000000000018103741477240017019},
+    {0x8d1d, -0.00000000000000000000000000000048379360203007364629},
+    {0x8d23, -0.00000000000000000000000000000050228252949619111048},
+    {0x8d30, -0.00000000000000000000000000000054234187233944561622},
+    {0x8dee, -0.00000000000000000000000000000146678824564531882569},
+    {0x8e54, -0.00000000000000000000000000000261310174854460160543},
+    {0x8e68, -0.00000000000000000000000000000285962078142616779462},
+    {0x8eb0, -0.00000000000000000000000000000433873497871556492976},
+    {0x8efe, -0.00000000000000000000000000000626158343519178120546},
+    {0x8f3d, -0.00000000000000000000000000000931841944292320195143},
+    {0x8f95, -0.00000000000000000000000000001469253435974134487579},
+    {0x8fa3, -0.00000000000000000000000000001607304094387811553526},
+    {0x8fbe, -0.00000000000000000000000000001873544649899903037853},
+    {0x9009, -0.00000000000000000000000000002701848600381965433535},
+    {0x9017, -0.00000000000000000000000000002977949917209319565429},
+    {0x90a8, -0.00000000000000000000000000006626431603856499165459},
+    {0x90d2, -0.00000000000000000000000000008283039504820623956823},
+    {0x9260, -0.00000000000000000000000000070681937107802657764891},
+    {0x92d0, -0.00000000000000000000000000131266454628776364420512},
+    {0x92ed, -0.00000000000000000000000000149568027629903838306064},
+    {0x9356, -0.00000000000000000000000000270105973947674442172976},
+    {0x94af, -0.00000000000000000000000001767048427695066444122272},
+    {0x9503, -0.00000000000000000000000002645523931749185190628773},
+    {0x953c, -0.00000000000000000000000003796629764647685617085567},
+    {0x9556, -0.00000000000000000000000004321695583162791074767614},
+    {0x9598, -0.00000000000000000000000006139231108792002274436236},
+    {0x9599, -0.00000000000000000000000006179620787139318078873317},
+    {0x95cf, -0.00000000000000000000000008360663417894371518475664},
+    {0x95df, -0.00000000000000000000000009006898271451424389468952},
+    {0x95f4, -0.00000000000000000000000009855081516745056282647643},
+    {0x960f, -0.00000000000000000000000011551448007332320069005024},
+    {0x9615, -0.00000000000000000000000012036124147500109722249990},
+    {0x966a, -0.00000000000000000000000018902369466543796476553675},
+    {0x968c, -0.00000000000000000000000022618219874496850484765081},
+    {0x96b3, -0.00000000000000000000000028919009696678115976949640},
+    {0x96d7, -0.00000000000000000000000034735123378691591815889232},
+    {0x9719, -0.00000000000000000000000049436966297114544630986535},
+    {0x9761, -0.00000000000000000000000072701421025168447986744905},
+    {0x97f0, -0.00000000000000000000000155096364853692689038389130},
+    {0x97fd, -0.00000000000000000000000163497417949934376361301874},
+    {0x983c, -0.00000000000000000000000242984304937451879493476303},
+    {0x987d, -0.00000000000000000000000326994835899868752722603749},
+    {0x9895, -0.00000000000000000000000385155972720003511111999672},
+    {0x98b3, -0.00000000000000000000000462704155146849855631194237},
+    {0x98fa, -0.00000000000000000000000646234853557052870993288041},
+    {0x998f, -0.00000000000000000000001478585344938536968832643037},
+    {0x999b, -0.00000000000000000000001602662436821491120063354341},
+    {0x99e4, -0.00000000000000000000002357464745776128873383514772},
+    {0x9a8b, -0.00000000000000000000005748905257243542340356290410},
+    {0x9a95, -0.00000000000000000000006162495563520056177791994756},
+    {0x9a9f, -0.00000000000000000000006576085869796570015227699102},
+    {0x9bb2, -0.00000000000000000000029447629806887785225422149438},
+    {0x9bc4, -0.00000000000000000000032425480012078684854959220729},
+    {0x9c09, -0.00000000000000000000045329497567905916582953196325},
+    {0x9c79, -0.00000000000000000000082387189010281556417192305730},
+    {0x9cca, -0.00000000000000000000133672386988569272259219644639},
+    {0x9d21, -0.00000000000000000000213081725793659929046874879077},
+    {0x9d2d, -0.00000000000000000000228963593554678060404405925965},
+    {0x9d53, -0.00000000000000000000279256174797902143036587574443},
+    {0x9d59, -0.00000000000000000000287197108678411208715353097887},
+    {0x9d9b, -0.00000000000000000000410281583826301726736218711267},
+    {0x9dc9, -0.00000000000000000000532042569994107400477290070739},
+    {0x9e3a, -0.00000000000000000000984675801183124144166924907040},
+    {0x9eb4, -0.00000000000000000001905824131322175762903725626529},
+    {0x9eb7, -0.00000000000000000001937587866844212025618787720305},
+    {0x9ec3, -0.00000000000000000002064642808932357076479036095407},
+    {0x9ee0, -0.00000000000000000002371692252312040949391303001903},
+    {0x9f0a, -0.00000000000000000002922263668027336169785712627345},
+    {0x9f5b, -0.00000000000000000004637505386217294356399065691221},
+    {0x9fb7, -0.00000000000000000007750351467376848102475150881219},
+    {0xa007, -0.00000000000000000011434944787933054577422353759175},
+    {0xa06c, -0.00000000000000000019989977555201488002012411016040},
+    {0xa07a, -0.00000000000000000021175823681357508476708062516991},
+    {0xa0c0, -0.00000000000000000032526065174565133020223584026098},
+    {0xa0e5, -0.00000000000000000038794108984246955529329170531128},
+    {0xa108, -0.00000000000000000046078592330633938445316744036973},
+    {0xa128, -0.00000000000000000056920614055488982785391272045672},
+    {0xa178, -0.00000000000000000084025668367626593635577592067420},
+    {0xa1f6, -0.00000000000000000166696084019646306728645868133754},
+    {0xa305, -0.00000000000000000720994444702860448614956112578511},
+    {0xa312, -0.00000000000000000791467585914418236825440544635057},
+    {0xa3f2, -0.00000000000000002623769257414920730298035778105259},
+    {0xa3f3, -0.00000000000000002634611279139775774638110306113958},
+    {0xa40a, -0.00000000000000002992397996059992237860569730401039},
+    {0xa412, -0.00000000000000003165870343657672947301762178540230},
+    {0xa413, -0.00000000000000003187554387107383035981911234557629},
+    {0xa421, -0.00000000000000003491130995403324277503998018801212},
+    {0xa450, -0.00000000000000004510281037539698445471003651618958},
+    {0xa456, -0.00000000000000004640385298237958977551897987723351},
+    {0xa4a4, -0.00000000000000007112366251504909087088890373706818},
+    {0xa4c9, -0.00000000000000008716985466783455649419920518994331},
+    {0xa4ed, -0.00000000000000010278236595162582034390652552247047},
+    {0xa522, -0.00000000000000014051260155412137464736588299274445},
+    {0xa53a, -0.00000000000000016132928326584305978030897676944733},
+    {0xa541, -0.0000000000000001674008154317618846107507124543190},
+    {0xa6f7, -0.00000000000000171390679426508540927898138761520386},
+    {0xa76c, -0.00000000000000327515792264421179424971342086791992},
+    {0xa7a3, -0.00000000000000452415882534751290222629904747009277},
+    {0xa7ae, -0.00000000000000482947015711943095084279775619506836},
+    {0xa7bc, -0.00000000000000521804821573823573999106884002685547},
+    {0xa837, -0.00000000000001015854067532018234487622976303100586},
+    {0xa83e, -0.00000000000001054711873393898713402450084686279297},
+    {0xa881, -0.00000000000001432187701766451937146484851837158203},
+    {0xa8b9, -0.00000000000002053912595556539599783718585968017578},
+    {0xa8c2, -0.00000000000002153832667772803688421845436096191406},
+    {0xa8ca, -0.00000000000002242650509742816211655735969543457031},
+    {0xa8d6, -0.00000000000002375877272697834996506571769714355469},
+    {0xa8e4, -0.00000000000002531308496145356912165880203247070312},
+    {0xa92d, -0.00000000000003841371665203041629865765571594238281},
+    {0xa933, -0.00000000000003974598428158060414716601371765136719},
+    {0xa937, -0.00000000000004063416270128072937950491905212402344},
+    {0xa950, -0.0000000000000461852778244065120816230773925781250},
+    {0xa956, -0.00000000000004751754545395669993013143539428710938},
+    {0xa9b8, -0.0000000000000817124146124115213751792907714843750},
+    {0xa9c6, -0.00000000000008792966355031239800155162811279296875},
+    {0xa9f0, -0.000000000000106581410364015027880668640136718750},
+    {0xaa2d, -0.00000000000015365486660812166519463062286376953125},
+    {0xaae6, -0.0000000000004085620730620576068758964538574218750},
+    {0xaaff, -0.00000000000045297099404706386849284172058105468750},
+    {0xab12, -0.000000000000518696197104873135685920715332031250},
+    {0xab42, -0.000000000000689226453687297180294990539550781250},
+    {0xabfa, -0.00000000000177635683940025046467781066894531250},
+    {0xac3a, -0.0000000000026432189770275726914405822753906250},
+    {0xadbb, -0.00000000002125943865394219756126403808593750},
+    {0xadc3, -0.00000000002216893335571512579917907714843750},
+    {0xadda, -0.0000000000247837306233122944831848144531250},
+    {0xaddf, -0.00000000002535216481192037463188171386718750},
+    {0xae16, -0.000000000034106051316484808921813964843750},
+    {0xae77, -0.0000000000561612978344783186912536621093750},
+    {0xaee2, -0.00000000010277290130034089088439941406250},
+    {0xb03e, -0.00000000069121597334742546081542968750},
+    {0xb050, -0.00000000075669959187507629394531250},
+    {0xb075, -0.000000000891304807737469673156738281250},
+    {0xb11d, -0.0000000022846506908535957336425781250},
+    {0xb125, -0.0000000024010660126805305480957031250},
+    {0xb139, -0.0000000026921043172478675842285156250},
+    {0xb155, -0.0000000030995579436421394348144531250},
+    {0xb18d, -0.000000004103640094399452209472656250},
+    {0xb23c, -0.000000010943040251731872558593750},
+    {0xb2a8, -0.0000000195577740669250488281250},
+    {0xb341, -0.000000044936314225196838378906250},
+    {0xb369, -0.000000054249539971351623535156250},
+    {0xb37b, -0.000000058440491557121276855468750},
+    {0xb3c6, -0.0000000922009348869323730468750},
+    {0xb3c9, -0.00000009359791874885559082031250},
+    {0xb3dc, -0.000000102445483207702636718750},
+    {0xb3e2, -0.0000001052394509315490722656250},
+    {0xb404, -0.00000012293457984924316406250},
+    {0xb42d, -0.0000001611188054084777832031250},
+    {0xb487, -0.000000251457095146179199218750},
+    {0xb499, -0.000000284984707832336425781250},
+    {0xb49b, -0.000000288709998130798339843750},
+    {0xb4be, -0.00000035390257835388183593750},
+    {0xb599, -0.0000011399388313293457031250},
+    {0xb5be, -0.000001415610313415527343750},
+    {0xb661, -0.000003352761268615722656250},
+    {0xb67f, -0.000003799796104431152343750},
+    {0xb6f4, -0.000007271766662597656250},
+    {0xb70f, -0.0000085234642028808593750},
+    {0xb729, -0.0000100731849670410156250},
+    {0xb731, -0.0000105500221252441406250},
+    {0xb735, -0.0000107884407043457031250},
+    {0xb76f, -0.0000142455101013183593750},
+    {0xb770, -0.000014305114746093750},
+    {0xb7a4, -0.0000195503234863281250},
+    {0xb7b1, -0.000021100044250488281250},
+    {0xb829, -0.00004029273986816406250},
+    {0xb882, -0.000061988830566406250},
+    {0xb9a6, -0.0003166198730468750},
+    {0xb9c5, -0.00037574768066406250},
+    {0xb9cc, -0.000389099121093750},
+    {0xb9d3, -0.00040245056152343750},
+    {0xb9dd, -0.00042152404785156250},
+    {0xbb04, -0.002014160156250},
+    {0xbb14, -0.002258300781250},
+    {0xbb19, -0.00233459472656250},
+    {0xbb33, -0.00273132324218750},
+    {0xbb66, -0.0035095214843750},
+    {0xbbc5, -0.0060119628906250},
+    {0xbc0d, -0.008605957031250},
+    {0xbcb0, -0.0214843750},
+    {0xbcc8, -0.02441406250},
+    {0xbce0, -0.027343750},
+    {0xbce8, -0.02832031250},
+    {0xbd06, -0.032714843750},
+    {0xbd77, -0.0603027343750},
+    {0xbe31, -0.17285156250},
+    {0xbe3a, -0.1816406250},
+    {0xbe5d, -0.21582031250},
+    {0xbe85, -0.2597656250},
+    {0xbe9a, -0.300781250},
+    {0xbea5, -0.3222656250},
+    {0xbeb0, -0.343750},
+    {0xbebf, -0.3730468750},
+    {0xbeee, -0.464843750},
+    {0xbf2b, -0.667968750},
+    {0xbfac, -1.343750},
+    {0xc022, -2.531250},
+    {0xc026, -2.593750},
+    {0xc05e, -3.468750},
+    {0xc07e, -3.968750},
+    {0xc07f, -3.9843750},
+    {0xc086, -4.18750},
+    {0xc0ae, -5.43750},
+    {0xc0c2, -6.06250},
+    {0xc0e6, -7.18750},
+    {0xc13e, -11.8750},
+    {0xc198, -19.0},
+    {0xc1be, -23.750},
+    {0xc1c1, -24.1250},
+    {0xc1eb, -29.3750},
+    {0xc225, -41.250},
+    {0xc276, -61.50},
+    {0xc27f, -63.750},
+    {0xc29f, -79.50},
+    {0xc313, -147.0},
+    {0xc31b, -155.0},
+    {0xc324, -164.0},
+    {0xc35b, -219.0},
+    {0xc394, -296.0},
+    {0xc39d, -314.0},
+    {0xc3b5, -362.0},
+    {0xc3be, -380.0},
+    {0xc429, -676.0},
+    {0xc444, -784.0},
+    {0xc44b, -812.0},
+    {0xc4b5, -1448.0},
+    {0xc4eb, -1880.0},
+    {0xc523, -2608.0},
+    {0xc557, -3440.0},
+    {0xc55e, -3552.0},
+    {0xc56d, -3792.0},
+    {0xc58b, -4448.0},
+    {0xc64d, -13120.0},
+    {0xc6b8, -23552.0},
+    {0xc6ca, -25856.0},
+    {0xc777, -63232.0},
+    {0xc7d6, -109568.0},
+    {0xc868, -237568.0},
+    {0xc8ca, -413696.0},
+    {0xc910, -589824.0},
+    {0xc9c5, -1613824.0},
+    {0xc9c8, -1638400.0},
+    {0xc9df, -1826816.0},
+    {0xca3a, -3047424.0},
+    {0xca42, -3178496.0},
+    {0xca6b, -3850240.0},
+    {0xcaa0, -5242880.0},
+    {0xcaa2, -5308416.0},
+    {0xcaac, -5636096.0},
+    {0xcb3a, -12189696.0},
+    {0xcb84, -17301504.0},
+    {0xcc50, -54525952.0},
+    {0xcc89, -71827456.0},
+    {0xcc94, -77594624.0},
+    {0xccaf, -91750400.0},
+    {0xcce0, -117440512.0},
+    {0xcce1, -117964800.0},
+    {0xcd6d, -248512512.0},
+    {0xcda8, -352321536.0},
+    {0xcdba, -390070272.0},
+    {0xcdd0, -436207616.0},
+    {0xcde5, -480247808.0},
+    {0xcdf7, -517996544.0},
+    {0xce30, -738197504.0},
+    {0xcec2, -1627389952.0},
+    {0xcf03, -2197815296.0},
+    {0xcf25, -2768240640.0},
+    {0xcf57, -3607101440.0},
+    {0xd036, -12213813248.0},
+    {0xd09e, -21206401024.0},
+    {0xd103, -35165044736.0},
+    {0xd104, -35433480192.0},
+    {0xd11f, -42681237504.0},
+    {0xd125, -44291850240.0},
+    {0xd19c, -83751862272.0},
+    {0xd1c7, -106837311488.0},
+    {0xd1cf, -111132278784.0},
+    {0xd1d8, -115964116992.0},
+    {0xd231, -190052302848.0},
+    {0xd28a, -296352743424.0},
+    {0xd294, -317827579904.0},
+    {0xd2be, -408021893120.0},
+    {0xd2c1, -414464344064.0},
+    {0xd2c6, -425201762304.0},
+    {0xd2db, -470298918912.0},
+    {0xd334, -773094113280.0},
+    {0xd36f, -1026497183744.0},
+    {0xd375, -1052266987520.0},
+    {0xd3c3, -1675037245440.0},
+    {0xd3d5, -1829656068096.0},
+    {0xd3f2, -2078764171264.0},
+    {0xd44c, -3504693313536.0},
+    {0xd49b, -5325759447040.0},
+    {0xd4cd, -7043746365440.0},
+    {0xd4e8, -7971459301376.0},
+    {0xd538, -12644383719424.0},
+    {0xd54c, -14018773254144.0},
+    {0xd554, -14568529068032.0},
+    {0xd5a3, -22402549415936.0},
+    {0xd5bf, -26250840113152.0},
+    {0xd64a, -55525337202688.0},
+    {0xd74f, -227598906949632.0},
+    {0xd75f, -245191092994048.0},
+    {0xd762, -248489627877376.0},
+    {0xd7d6, -470590976688128.0},
+    {0xd7db, -481586092965888.0},
+    {0xd819, -672901116198912.0},
+    {0xd82d, -760862046420992.0},
+    {0xd85b, -963172185931776.0},
+    {0xd936, -3201777860083712.0},
+    {0xd967, -4063794976260096.0},
+    {0xd976, -4327677766926336.0},
+    {0xd985, -4679521487814656.0},
+    {0xd98c, -4925812092436480.0},
+    {0xd9c9, -7072058789855232.0},
+    {0xda1b, -10907155347537920.0},
+    {0xda5a, -15340386230730752.0},
+    {0xdab6, -25614222880669696.0},
+    {0xdacc, -28710447624486912.0},
+    {0xdb00, -36028797018963968.0},
+    {0xdb0d, -39687971716202496.0},
+    {0xdb24, -46161896180547584.0},
+    {0xdb4c, -57420895248973824.0},
+    {0xdbbd, -106397541196627968.0},
+    {0xdbbf, -107523441103470592.0},
+    {0xdbcc, -114841790497947648.0},
+    {0xdbf4, -137359788634800128.0},
+    {0xdc88, -306244774661193728.0},
+    {0xdc9c, -351280770934898688.0},
+    {0xdd24, -738590338888761344.0},
+    {0xdd71, -1085367510196289536.0},
+    {0xdd86, -1206964700135292928.0},
+    {0xdd8f, -1288029493427961856.0},
+    {0xdea9, -6088866696204910592.0},
+    {0xded6, -7710162562058289152.0},
+    {0xdedf, -8034421735228964864.0},
+    {0xdef6, -8863084066665136128.0},
+    {0xdf62, -16285016252571713536.0},
+    {0xdf6a, -16861477004875137024.0},
+    {0xdf71, -17365880163140632576.0},
+    {0xdfc9, -28967152803247030272.0},
+    {0xdff0, -34587645138205409280.0},
+    {0xdff5, -35308221078584688640.0},
+    {0xe037, -52746158835763249152.0},
+    {0xe07e, -73210515542534782976.0},
+    {0xe09a, -88774955854727217152.0},
+    {0xe09d, -90504338111637487616.0},
+    {0xe0cc, -117597993469898391552.0},
+    {0xe113, -169479461177206505472.0},
+    {0xe18f, -329735550317558235136.0},
+    {0xe1aa, -391993311566327971840.0},
+    {0xe21a, -710199646837817737216.0},
+    {0xe233, -825491797298502434816.0},
+    {0xe238, -848550227390639374336.0},
+    {0xe247, -917725517667050192896.0},
+    {0xe262, -1042241040164589666304.0},
+    {0xe26a, -1079134528312008769536.0},
+    {0xe271, -1111416330441000484864.0},
+    {0xe28b, -1282048713122813837312.0},
+    {0xe290, -1328165573307087716352.0},
+    {0xe2b2, -1641760222560150093824.0},
+    {0xe2eb, -2167492428660872314880.0},
+    {0xe2f5, -2259726149029420072960.0},
+    {0xe34b, -3744689046963038978048.0},
+    {0xe363, -4187410904732068216832.0},
+    {0xe388, -5017514388048998039552.0},
+    {0xe3a0, -5902958103587056517120.0},
+    {0xe3bd, -6972869259862210510848.0},
+    {0xe3d6, -7895206463547688091648.0},
+    {0xe3e4, -8411715297611555536896.0},
+    {0xe406, -9887454823508319666176.0},
+    {0xe5a3, -96218217088469021229056.0},
+    {0xe5ae, -102711471002414783397888.0},
+    {0xe5d7, -126913599227121715118080.0},
+    {0xe5d9, -128094190847839126421504.0},
+    {0xe644, -231395957660612615471104.0},
+    {0xe66e, -280980805730743890214912.0},
+    {0xe6b8, -434457716424007359660032.0},
+    {0xe7f4, -2304514843640386864283648.0},
+    {0xe824, -3097872412762487260184576.0},
+    {0xe827, -3154540810556923002748928.0},
+    {0xe880, -4835703278458516698824704.0},
+    {0xe8f2, -9142501510835633133715456.0},
+    {0xe928, -12693721105953606334414848.0},
+    {0xe980, -19342813113834066795298816.0},
+    {0xe9d1, -31583187037432187189198848.0},
+    {0xe9fa, -37778931862957161709568000.0},
+    {0xea15, -45032486780644936757805056.0},
+    {0xea77, -74651169361203351538106368.0},
+    {0xeaca, -122101507781077546645323776.0},
+    {0xeacd, -123914896510499490407383040.0},
+    {0xead9, -131168451428187265455620096.0},
+    {0xeada, -131772914337994580042973184.0},
+    {0xeb79, -301022529084042664501837824.0},
+    {0xeba3, -394109817194369110954213376.0},
+    {0xebf3, -587537948332709778907201536.0},
+    {0xec09, -662491349148816787738984448.0},
+    {0xec13, -710848381933401954727231488.0},
+    {0xec48, -967140655691703339764940800.0},
+    {0xece9, -2253437727761668781652312064.0},
+    {0xed1f, -3075507285099616620452511744.0},
+    {0xed24, -3172221350668786954429005824.0},
+    {0xed9f, -6151014570199233240905023488.0},
+    {0xede6, -8897694032363670725837455360.0},
+    {0xedfe, -9826149061827705932011798528.0},
+    {0xeea3, -25223028300439623101069656064.0},
+    {0xeea5, -25532513310260968169794437120.0},
+    {0xeec4, -30329530962491816735028543488.0},
+    {0xeee0, -34662321099990647697175478272.0},
+    {0xef23, -50446056600879246202139312128.0},
+    {0xefa0, -99035203142830421991929937920.0},
+    {0xefab, -105843873358900013503875121152.0},
+    {0xf020, -198070406285660843983859875840.0},
+    {0xf099, -378809652021326364119132012544.0},
+    {0xf0a9, -418423733278458532915903987712.0},
+    {0xf0cc, -505079536028435152158842683392.0},
+    {0xf0e6, -569452418071274926453597143040.0},
+    {0xf16b, -1163663636928257458405176770560.0},
+    {0xf19b, -1535045648713871540874914037760.0},
+    {0xf1bc, -1861861819085211933448282832896.0},
+    {0xf1da, -2158967428513703199424072646656.0},
+    {0xf231, -3505846191256196938514319802368.0},
+    {0xf253, -4179285572627443808059443380224.0},
+    {0xf287, -5347900969712842787564216647680.0},
+    {0xf296, -5942112188569825319515796275200.0},
+    {0xf2f3, -9626221745483117017615589965824.0},
+    {0xf339, -14657210065138902454805630812160.0},
+    {0xf36e, -18856302678394912347263460179968.0},
+    {0xf3b3, -28363682180106632858488734220288.0},
+    {0xf3c0, -30423614405477505635920876929024.0},
+    {0xf3da, -34543478856219251190785162346496.0},
+    {0xf3eb, -37237236381704238668965656657920.0},
+    {0xf407, -42783207757702742300513733181440.0},
+    {0xf418, -48170722808672717256874721804288.0},
+    {0xf425, -52290587259414462811739007221760.0},
+    {0xf473, -77009773963864936140924719726592.0},
+    {0xf515, -188879939434006180823008777601024.0},
+    {0xf57a, -316912650057057350374175801344000.0},
+    {0xf594, -375224577667555902843024148791296.0},
+    {0xf61c, -791013974542415146533942800154624.0},
+    {0xf634, -912708432164325169077626307870720.0},
+    {0xf69f, -1612451563490307798703806477238272.0},
+    {0xf6b2, -1805134454724998667731305364455424.0},
+    {0xf6f0, -2433889152438200450873670154321920.0},
+    {0xf71e, -3204620717376963926983665703190528.0},
+    {0xf735, -3671116138260952346734452482768896.0},
+    {0xf748, -4056481920730334084789450257203200.0},
+    {0xf75f, -4522977341614322504540237036781568.0},
+    {0xf796, -6084722881095501127184175385804800.0},
+    {0xf8a3, -26448262123161778232827215676964864.0},
+    {0xf8d0, -33749929580476379585448226139930624.0},
+    {0xf8ef, -38779967162181993850587144458862592.0},
+    {0xf9a3, -105793048492647112931308862707859456.0},
+    {0xfa0d, -183028464263352673905699995605008384.0},
+    {0xfa18, -197307280624323449884158860510363648.0},
+    {0xfa31, -229759135990166122562474462567989248.0},
+    {0xfaa9, -438749084546192934610826939819098112.0},
+    {0xfab5, -469902865697401900382009917794418688.0},
+    {0xfacd, -532210427999819831924375873745059840.0},
+    {0xfb84, -1370766370653194493932051030914105344.0},
+    {0xfba8, -1744611744467702083186246766617952256.0},
+    {0xfbfa, -2596148429267413814265248164610048000.0},
+    {0xfc20, -3323069989462289682259517650700861440.0},
+    {0xfc76, -5109220108798270386474008387952574464.0},
+    {0xfc93, -6106141105636957291151863683162832896.0},
+    {0xfccf, -8598443597733674552846501921188478976.0},
+    {0xfcdc, -9138442471021296626213673539427368960.0},
+    {0xfcf0, -9969209968386869046778552952102584320.0},
+    {0xfd5b, -18193808192306036010370859137587216384.0},
+    {0xfda9, -28079941410956347815092924148422279168.0},
+    {0xfdca, -33563006893569125790821128272078700544.0},
+    {0xfe1d, -52172198834557948011474427116003524608.0},
+    {0xfe24, -54498347827181550789056089471494127616.0},
+    {0xfe4b, -67458320786084480549868208309227487232.0},
+    {0xfe4f, -68787548781869396422772015369507831808.0},
+    {0xfe64, -75765995759740204755517002435979640832.0},
+    {0xfea2, -107667467658578185705208371882707910656.0},
+    {0xfec2, -128935115591136839671669284847193423872.0},
+    {0xfee3, -150867377521587951574582101341819109376.0},
+    {0xfee9, -154855061508942699193293522522660143104.0},
+    {0xff57, -285784019093756912674318517960274083840.0},
+    {0xff60, -297747071055821155530452781502797185024.0},
+    {0xff8e, std::numeric_limits<float>::quiet_NaN()},
+    {0xffb0, std::numeric_limits<float>::quiet_NaN()},
+    {0xfffa, std::numeric_limits<float>::quiet_NaN()},
 };
 
-// TEST_CASE(clang_bf16) {
-
-//     std::set<uint16_t> samples = {0};
-//     auto ns                    = migraphx::range(1, 65535);
-//     std::sample(ns.begin(),
-//                 ns.end(),
-//                 std::inserter(samples, samples.begin()),
-//                 1022,
-//                 std::mt19937{std::random_device{}()});
-//     samples.insert(65536);
-
-
-//     for(uint16_t i : samples)
-//     {
-
-//         float f = migraphx::bit_cast<__bf16>(i);
-
-//         std::cout << "{0x" << std::hex << std::setw(4) << std::setfill('0') << i  << ", ";
-
-//         if(std::isfinite(f))
-//         {
-//             std::cout << std::fixed << std::setprecision(40) << static_cast<float>(f);
-//         }
-//         else if(std::isinf(f))
-//         {
-//             if(f < 0)
-//                 std::cout << "-";
-//             std::cout << "std::numeric_limits<float>::infinity()";
-//         }
-//         else if(std::isnan(f))
-//         {
-//             std::cout << "std::numeric_limits<float>::quiet_NaN()";
-//         }
-//         std::cout << "},\n";
-//     }
-
-// }
-
-
-TEST_CASE(check_half_values)
+TEST_CASE(check_bf16_values)
 {
-    for(auto [x, f] : half_lut)
+    for(auto [x, f] : bf16_lut)
     {
 
         auto h = migraphx::bit_cast<migraphx::bf16>(x);
@@ -1261,7 +1212,7 @@ TEST_CASE(test_binary_ops)
     auto c = migraphx::bf16(0.0);
     auto d = migraphx::bf16(-0.0);
     EXPECT(migraphx::float_equal((c + d), c));
-    // EXPECT(migraphx::float_equal((c + d), d));
+    EXPECT(migraphx::float_equal((c + d), d));
     EXPECT(migraphx::float_equal((a + b), c));
     EXPECT(migraphx::float_equal((a + b), d));
